% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions}
\alias{actions}
\alias{use_github_actions}
\alias{use_github_actions_badge}
\alias{use_action}
\alias{use_action_ci_quick}
\alias{use_action_ci_full}
\alias{use_action_pr_commands}
\title{GitHub Actions setup and badges}
\usage{
use_github_actions()

use_github_actions_badge(name = "R")

use_action(name,
  url = glue(" https://raw.githubusercontent.com/r-lib/actions/master/examples/{name}"),
  save_as = basename(url), ignore = TRUE, open = FALSE)

use_action_ci_quick(save_as = "R.yaml", ignore = TRUE, open = FALSE)

use_action_ci_full(save_as = "R.yaml", ignore = TRUE, open = FALSE)

use_action_pr_commands(save_as = "pr-commands.yaml", ignore = true,
  open = FALSE)
}
\arguments{
\item{name}{The name to give to the GitHub Actions workflow}

\item{url}{The full URL to the action yaml file.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{ignore}{Should the newly created file be added to \code{.Rbuildignore}?}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Sets up continous integration (CI) for an R package that is developed on
GitHub using \href{https://github.com/features/actions}{GitHub Actions}. These functions
\itemize{
\item Add the necessary configuration files and place them in \code{.Rbuildignore}.
\item Provide the markdown to insert a badge into your README
}
}
\section{\code{use_github_actions()}}{

Adds a basic ‘r.yaml’ file to the \code{.github/workflows} directory of a
package. This is a configuration file for the GitHub Actions service'
}

\section{\code{use_github_actions_badge()}}{

Only adds the \href{https://github.com/features/actions}{GitHub Actions} badge. Use for a project
where GitHub actions is already configured.
}

\section{\code{use_action()}}{

Use a specific action, either one of the example actions from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions/examples} or a custom action
given by the \code{url} parameter.
}

