% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_github}
\alias{browse_github}
\alias{browse_github_issues}
\alias{browse_github_pulls}
\alias{browse_travis}
\alias{browse_cran}
\title{Quickly browse to important package webpages}
\usage{
browse_github(package = NULL)

browse_github_issues(package = NULL, number = NULL)

browse_github_pulls(package = NULL, number = NULL)

browse_travis(package = NULL)

browse_cran(package = NULL)
}
\arguments{
\item{package}{Name of package; leave as \code{NULL} to use current package}

\item{number}{For GitHub issues and pull requests. Can be a number or
\code{"new"}.}
}
\description{
Quickly browse to important package webpages
}
\examples{
browse_cran("MASS")
}
