% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_rseek}
\alias{browse_rseek}
\title{Search the web with rseek}
\usage{
browse_rseek(term)
}
\arguments{
\item{term}{Search term}
}
\description{
This function takes you to a search page on rseek.org and returns the target
URL invisibly.
}
\examples{
browse_rseek("How to use usethis")
}
