% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{create_air_toml}
\alias{create_air_toml}
\title{Creates an empty \code{air.toml}}
\usage{
create_air_toml(ignore = FALSE, dot_prefix = FALSE)
}
\description{
If either \code{air.toml} or \code{.air.toml} already exist, no new file is created.
}
\keyword{internal}
