% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_from_github}
\alias{create_from_github}
\title{Create a project from a github repository}
\usage{
create_from_github(repo, path, fork = NA, open = TRUE)
}
\arguments{
\item{repo}{Full name of repository: \code{owner/repo}}

\item{path}{A path. If it exists, it will be used. If it does not exist, it
will be created (providing that the parent path exists).}

\item{fork}{Create a fork before cloning? Defaults to \code{TRUE} if you can't
push to \code{repo}, \code{FALSE} if you can.}

\item{open}{If \code{TRUE} and in RStudio, new project will be opened in a new
instance, if possible, or will be switched to, otherwise. If \code{TRUE} and not
in RStudio, working directory will be set to the new project.}
}
\description{
Create a project from a github repository
}
