% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating-files.R, R/r.R, R/test.R
\name{creating-files}
\alias{creating-files}
\alias{use_r}
\alias{use_test}
\title{Create or edit files}
\usage{
use_r(name = NULL)

use_test(name = NULL, open = interactive())
}
\arguments{
\item{name}{\itemize{
\item for \code{use_r()}: File name, without extension; will create if it doesn't already
exist. If not specified, and you're currently in a test file, will guess
name based on test name.
\item for \code{use_test()}: Base of test file name. If \code{NULL}, and you're using RStudio, will
be based on the name of the file open in the source editor.
}}

\item{open}{\itemize{
\item If TRUE, opens the file for editing. Defaults to return value of \code{interactive()}.
}}
}
\description{
There are two functions for creating files:
\itemize{
\item \code{use_r()} creates \code{<name>.R} file in the \code{R} directory.
\item \code{use_test()} sets up individual test files: creates \code{tests/testthat/test-<name>.R} and, optionally, opens it for editing.
}
}
\seealso{
\code{\link[=use_testthat]{use_testthat()}} to set up the testing infrastructure. The \href{https://r-pkgs.org/tests.html}{testing}
and \href{https://r-pkgs.org/r.html}{R code} chapters of \href{https://r-pkgs.org}{R Packages}.
}
