% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{download_zip}
\alias{download_zip}
\title{Download a ZIP file}
\usage{
download_zip(url, destdir = NULL, pedantic = TRUE)
}
\arguments{
\item{url}{Download link for the ZIP file, possibly behind a shortlink or
other redirect. See Details.}

\item{destdir}{Path to existing local directory where the ZIP file will be
stored. Defaults to current working directory.}

\item{pedantic}{Logical. When \code{TRUE} (default) and \code{destdir = NULL} and in an
interactive session, the user is told where the ZIP file will be stored. If
happy, user can elect to proceed. Otherwise, user can abort and try again
with the desired \code{destdir}. Intentional friction.}
}
\value{
Path to downloaded ZIP file
}
\description{
Special-purpose function to download a ZIP file and automatically determine
the file name, which often determines the folder name after unpacking.
Developed for use in live teaching, with DropBox and GitHub as primary
targets, possibly via shortlinks. Both platforms offer a way to download an
entire folder or repo as a ZIP file, with information about the original
folder or repo transmitted in the \code{Content-Disposition} header. In the
absence of this header, a filename is generated from the input URL. In either
case, the filename is sanitized.
}
\section{DropBox}{


To make a folder available for ZIP download, create a shared link for it:
\itemize{
\item \url{https://www.dropbox.com/help/files-folders/view-only-access}
}

A shared link will have this form:\preformatted{https://www.dropbox.com/sh/12345abcde/6789wxyz?dl=0
}

Replace the \code{dl=0} at the end with \code{dl=1} to create a download link. The ZIP
download link will have this form:\preformatted{https://www.dropbox.com/sh/12345abcde/6789wxyz?dl=1
}

This download link (or a shortlink that points to it) is suitable as input
for \code{download_zip()}. After one or more redirections, this link will
eventually lead to a download URL. For more details, see
\url{https://www.dropbox.com/help/desktop-web/force-download} and
\url{https://www.dropbox.com/en/help/desktop-web/download-entire-folders}.
}

\section{GitHub}{


Click on the repo's "Clone or download" button, to reveal a "Download ZIP"
button. Capture this URL, which will have this form:\preformatted{https://github.com/r-lib/usethis/archive/master.zip
}

This download link (or a shortlink that points to it) is suitable as input
for \code{download_zip()}. After one or more redirections, this link will
eventually lead to a download URL. Here's an alternative link that also leads
to ZIP download, albeit with a different filenaming scheme:\preformatted{http://github.com/r-lib/usethis/zipball/master/
}
}

\examples{
\dontrun{
download_zip("http://bit.ly/uusseetthhiiss")
}
}
\seealso{
Other download functions: \code{\link{tidy_unzip}},
  \code{\link{use_course}}
}
\concept{download functions}
\keyword{internal}
