% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-fork.R
\name{gh_fork_status}
\alias{gh_fork_status}
\alias{gh_fork_cleanup}
\title{Find forks with no open pull requests}
\usage{
gh_fork_status()

gh_fork_cleanup(api_key)
}
\arguments{
\item{api_key}{An API key with \code{delete_repo} scope. We recommend making
this a very shortlived token.}
}
\description{
\code{gh_fork_status()} reports all forks that the current user has, along with
the number of open PRs in each.

\code{gh_fork_cleanup()} deletes all forks with zero open PRs.
}
\examples{
\dontrun{
# Find and optionally delete forks with no open PRs
gh_fork_status()
gh_fork_cleanup()
}
}
