% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_credentials}
\alias{git_credentials}
\title{Produce or register git credentials}
\usage{
git_credentials(protocol = git_protocol(), auth_token = github_token())
}
\arguments{
\item{protocol}{Optional. Should be "ssh" or "https", if specified. Defaults
to the option \code{usethis.protocol} and, if unset, to an interactive choice
or, in non-interactive sessions, "ssh". \code{NA} triggers the interactive menu.}

\item{auth_token}{GitHub personal access token (PAT).}
}
\value{
Either \code{NULL} or a git2r credential object, invisibly, i.e.
something to be passed to git2r as \code{credentials}.
}
\description{
\emph{This is not needed any more and will be removed soon.}
}
\examples{
git_credentials()
git_credentials(protocol = "ssh")
\dontrun{
# these calls look for a GitHub PAT
git_credentials(protocol = "https")
git_credentials(protocol = "https", auth_token = "MY_GITHUB_PAT")
}
}
