% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_vaccinate}
\alias{git_vaccinate}
\title{Vaccinate your global git ignore}
\usage{
git_vaccinate()
}
\description{
Adds \code{.DS_Store}, \code{.Rproj.user}, and \code{.Rhistory} to your global
\code{.gitignore}. This is a good practice as it decreases the chance that you
accidentally leak credentials to GitHub.
}
