% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{knit_rmd}
\alias{knit_rmd}
\title{Knit a \code{.Rmd} in a clean session}
\usage{
knit_rmd(path, ..., quiet = FALSE)
}
\arguments{
\item{path}{Path to file within current project}

\item{...}{Additional arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}}

\item{quiet}{If \code{TRUE}, will suppress output from knitr.}
}
\description{
\code{knit_rmd()} runs a \code{\link[rmarkdown:render]{rmarkdown::render()}} in separate R session so that
the document doesn't inherit anything from your current session. It's
useful for rebuilding \code{README.Rmd} or \code{INDEX.Rmd}.
}
