% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license.R
\name{licenses}
\alias{use_gpl3_license}
\alias{use_mit_license}
\title{Add licenses}
\usage{
use_mit_license(copyright_holder = getOption("devtools.name", "<Author>"),
  base_path = ".")

use_gpl3_license(base_path = ".")
}
\arguments{
\item{copyright_holder}{The copyright holder for this package. Defaults to
\code{getOption("devtools.name")}.}

\item{base_path}{Path to package root.}
}
\description{
Adds the necessary infrastructure to declare your package as
distributed under either the MIT license (including the \code{LICENSE}
file), or GPL v3.
}

