% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{list_datasets}
\alias{list_datasets}
\title{List the datasets in a package}
\usage{
list_datasets(pkg)
}
\arguments{
\item{pkg}{A string naming a package.}
}
\value{
A data frame with two column, with one row per dataset.
\describe{
\item{Item}{Character. The name of the dataset.}
\item{Title}{Character. A sentence describing the dataset.}
}
}
\description{
A convenience wrapper for \code{\link[utils]{data}} to list the name and
title of all the datasets in an installed package.
}
\seealso{
\code{\link[utils]{data}}, \code{\link{use_readme_rmd}}
}
