% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr_init}
\alias{pr_init}
\alias{pr_fetch}
\alias{pr_push}
\alias{pr_pull}
\alias{pr_pull_source}
\alias{pr_sync}
\alias{pr_view}
\alias{pr_pause}
\alias{pr_finish}
\title{Helpers for GitHub pull requests}
\usage{
pr_init(branch)

pr_fetch(number, owner = NULL)

pr_push()

pr_pull()

pr_pull_source()

pr_sync()

pr_view()

pr_pause()

pr_finish()
}
\arguments{
\item{branch}{branch name. Should usually consist of lower case letters,
numbers, and \code{-}.}

\item{number}{Number of PR to fetch.}

\item{owner}{Name of the owner of the repository that is the target of the
pull request. Default of \code{NULL} uses the source repo.}
}
\description{
\itemize{
\item \code{pr_init("name")} creates and checks out a new local branch, in
anticipation of making a PR.
\item \code{pr_fetch(number)} downloads a remote PR so you can edit locally.
\item \code{pr_push()} pushes local changes to GitHub, after checking that there
aren't any remote changes you need to retrieve first. It will create
a PR if needed
\item \code{pr_pull()} retrieves changes from the GitHub PR. Run this if others
have made suggestions or pushed into your PR.
\item \code{pr_pull_source()} updates your PR with changes from the source
repository
\item \code{pr_sync()} is a shortcut for \code{pr_pull()}, \code{pr_pull_source()}, and
\code{pr_push()}
\item \code{pr_view()} opens the PR in the browser
\item \code{pr_pause()} makes sure you're synched with the PR and then switches
back to master.
\item \code{pr_finish()} changes local branch to \code{master}, pulls new changes, and
deletes the PR branch. Run this from the PR branch after is has been
merged.
}
}
\details{
These functions have been designed to support the git and GitHub best
practices described in \url{http://happygitwithr.com/}.
}
\examples{
\dontrun{
## scenario: current project is a local copy of fork of a repo owned by
## 'tidyverse', not you
pr_fetch(123, owner = "tidyverse")
}
}
