% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{pr_pull_upstream}
\alias{pr_pull_upstream}
\alias{pr_sync}
\title{Defunct PR functions}
\usage{
pr_pull_upstream()

pr_sync()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
\itemize{
\item \code{pr_pull_upstream()} has been replaced by \code{\link[=pr_merge_main]{pr_merge_main()}}.
\item \code{pr_sync()} has been replaced by \code{\link[=pr_pull]{pr_pull()}} + \code{\link[=pr_merge_main]{pr_merge_main()}} + \code{\link[=pr_push]{pr_push()}}
}
}
\keyword{internal}
