% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_activate}
\alias{proj_activate}
\title{Activate a project}
\usage{
proj_activate(path)
}
\arguments{
\item{path}{Project directory}
}
\value{
Single logical value indicating if current session is modified.
}
\description{
Activates a project in the usethis, R session, and (if relevant) RStudio or
Positron senses. If you are in RStudio or Positron, this will open a new
session or window. If not, it will change the working directory and
\link[=proj_set]{active project}.
}
\details{
\itemize{
\item If using RStudio desktop, the project is opened in a new session.
\item If using Positron, the project is opened in a new window.
\item If using RStudio or Positron on a server, the project is opened in a new
browser tab.
\item Otherwise, the working directory and active project are changed in the
current R session.
}
}
