% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_get}
\alias{proj_get}
\alias{proj_set}
\title{Get and set currently active project}
\usage{
proj_get()

proj_set(path = ".", force = FALSE)
}
\arguments{
\item{path}{Path to set.}

\item{force}{If \code{TRUE}, uses this path without checking if any parent
directories are existing projects.}
}
\description{
When attached, usethis uses rprojroot to find the project root of the
current working directory. It establishes the project root by looking for
for a \code{.here} file, an RStudio project, a package \code{DESCRIPTION}, a
\code{remake.yml}, or a \code{.projectile} file. It then stores the project directory
for use for the remainder of the session. If needed, you can manually
override by running \code{proj_set()}.
}
\keyword{internal}
