% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-lib.R
\name{r-lib}
\alias{r-lib}
\alias{use_lifecycle}
\title{Helpers from the r-lib organisation}
\usage{
use_lifecycle()
}
\description{
Helpers from the r-lib organisation
}
\details{
\itemize{
\item \code{use_lifecycle()}: adds a dependency on the lifecycle
package and imports the lifecycle SVG badges to the \code{man/figures}
folder. See \code{\link[lifecycle:badge]{lifecycle::badge()}} for how to incorporate these
badges in your documentation.
}
}
