% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme_to_vignette.R
\name{README_to_vignette}
\alias{README_to_vignette}
\title{Convert an existing \code{README.Rmd} into a vignette}
\usage{
README_to_vignette(vignette_aut = NULL, vignette_title = NULL,
  vignette_slug = "README-vignette", overwrite = FALSE)
}
\arguments{
\item{vignette_aut}{The intended author of the README/vignette. If not
provided, an attempt will be made to extract it from the package's
\code{DESCRIPTION} file (\code{Author} or \code{Authors@R}) as the first person listed
with an \code{aut} (author) role (if used).}

\item{vignette_title}{The intended title of the vignette. If not provided,
this will be taken from the Title field of the \code{DESCRIPTION} file.}

\item{vignette_slug}{filename to be used as the vignette. By default this
will be README-vignette. This will be converted to a compatible name if
required.}

\item{overwrite}{(logical) if the vignette already exists, should it be
overwritten?}
}
\description{
While many package lack vignettes, they often contain useful information in
the form of a \code{README.Rmd} file which could in theory form a first draft of a
vignette. This function attempts to convert the \code{README.Rmd} file into a
passable vignette, leveraging the same internals as \code{\link[=use_vignette]{use_vignette()}}. Once
created, an attempt is made to open the file for further editing.
}
\details{
This is not guaranteed to create a complete vignette ready to be knit, it
just does the easy part. Relative file paths will still need to be updated.

In order to perform the conversion from outside of the active project use
\code{\link[=proj_set]{proj_set()}} first to set the active project.
}
\examples{
\dontrun{
## within an existing package project context
README_to_vignette()}
}
