% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{rprofile-helper}
\alias{rprofile-helper}
\alias{use_usethis}
\alias{use_partial_warnings}
\title{Helpers to make useful changes to \code{.Rprofile}}
\usage{
use_usethis()

use_partial_warnings()
}
\description{
All functions open your \code{.Rprofile} and gives you the code you need to
paste in.
\itemize{
\item \code{use_usethis()}: makes usethis available in interactive sessions.
\item \code{use_partial_warning()}: warn on partial matches.
}
}
