% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-deprecated.R
\name{tidy-deprecated}
\alias{tidy-deprecated}
\alias{use_tidy_eval}
\alias{use_tidy_style}
\title{Deprecated tidyverse functions}
\usage{
use_tidy_eval()

use_tidy_style(strict = TRUE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item \code{use_tidy_eval()} is deprecated because there's no longer a need to
systematically import and re-export a large number of functions in order
to use tidy evaluation. Instead, use \code{\link[=use_import_from]{use_import_from()}} to tactically
import functions as you need them.
\item \code{use_tidy_style()} is deprecated because tidyverse packages are moving
towards the use of \href{https://posit-dev.github.io/air/}{Air} for formatting.
See \code{\link[=use_air]{use_air()}} for how to start using Air. To continue using the styler
package, see \code{styler::style_pkg()} and \code{styler::style_dir()}.
}
}
\keyword{internal}
