% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui-questions}
\alias{ui-questions}
\alias{ui_yeah}
\alias{ui_nope}
\alias{ui_yes_options}
\alias{ui_no_options}
\title{User interface - Questions}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 6.
}
\usage{
ui_yeah(
  x,
  yes = ui_yes_options,
  no = ui_no_options,
  n_yes = 1,
  n_no = 2,
  shuffle = TRUE,
  .envir = parent.frame()
)

ui_nope(
  x,
  yes = ui_yes_options,
  no = ui_no_options,
  n_yes = 1,
  n_no = 2,
  shuffle = TRUE,
  .envir = parent.frame()
)

ui_yes_options

ui_no_options
}
\arguments{
\item{x}{A character vector.

For block styles, conditions, and questions, each element of the
vector becomes a line, and the result is processed by \code{\link[glue:glue]{glue::glue()}}.
For inline styles, each element of the vector becomes an entry in a
comma separated list.}

\item{yes}{A character vector of "yes" strings, which are randomly sampled
from \code{ui_yes_options} to populate the menu.}

\item{no}{A character vector of "no" strings, which are randomly sampled to
from \code{ui_yes_options} to populate the menu.}

\item{n_yes}{An integer. The number of "yes" strings to include.}

\item{n_no}{An integer. The number of "no" strings to include.}

\item{shuffle}{A logical. Should the order of the menu options be randomly
shuffled?}

\item{.envir}{Used to ensure that \code{\link[glue:glue]{glue::glue()}} gets the correct
environment. For expert use only.}
}
\value{
A logical. \code{ui_yeah()} returns \code{TRUE} when the user selects a "yes"
option and \code{FALSE} otherwise, i.e. when user selects a "no" option or
refuses to make a selection (cancels). \code{ui_nope()} is the logical opposite
of \code{ui_yeah()}.
}
\description{
These functions are used to interact with the user by posing a simple yes or
no question. For details on the other \verb{ui_*()} functions, see the \link{ui} help
page.
}
\examples{
\dontrun{
ui_yeah("Do you like R?")
ui_nope("Have you tried turning it off and on again?", n_yes = 1, n_no = 1)
ui_yeah("Are you sure its plugged in?", yes = "Yes", no = "No", shuffle = FALSE)
}

# print yes and no options
ui_yes_options
ui_no_options
}
\seealso{
Other user interface functions: 
\code{\link{ui}}
}
\concept{user interface functions}
\keyword{datasets}
\keyword{internal}
