% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_}
\alias{use_}
\title{Create functions for inserting code into non-templated files.}
\usage{
use_(edit_function, todo_text, code, .return)
}
\arguments{
\item{edit_function}{function A function that opens the desired file for
editing. E.g. \link{edit_r_profile}}

\item{todo_text}{character Message describing the reasons for editing the
file.}

\item{code}{character The code to be added to the file. This will be
displayed, and also copied to the user's clipboard}

\item{.return}{Expression to evaluate and return invisibly}
}
\value{
A function.
}
\description{
Create functions for inserting code into non-templated files.
}
