% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{use_addin}
\alias{use_addin}
\title{Add minimal RStudio Addin binding}
\usage{
use_addin(addin = "new_addin", open = rlang::is_interactive())
}
\arguments{
\item{addin}{Name of the addin function, which should be defined in the
\code{R} folder.}

\item{open}{Open the newly created file for editing? Happens in RStudio or
Positron, if applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
This function helps you add a minimal
\href{https://rstudio.github.io/rstudioaddins/}{RStudio Addin} binding to
\code{inst/rstudio/addins.dcf}.
}
