% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{use_air}
\alias{use_air}
\title{Configure a project to use Air}
\usage{
use_air(vscode = TRUE, dot_prefix = FALSE)
}
\arguments{
\item{vscode}{Either:
\itemize{
\item \code{TRUE} to set up VS Code and Positron specific Air settings. This is the
default.
\item \code{FALSE} to opt out of those settings.
}}

\item{dot_prefix}{Either:
\itemize{
\item \code{FALSE} to create a configuration file named \code{air.toml}. This is the
default.
\item \code{TRUE} to create a configuration file named \code{.air.toml}.
}}
}
\description{
\href{https://github.com/posit-dev/air}{Air} is an extremely fast R code
formatter. This function sets up a project to use Air. Specifically, it:
\itemize{
\item Creates an empty \code{air.toml} configuration file. If either an \code{air.toml} or
\code{.air.toml} file already existed, nothing is changed. If the project is an
R package, \code{.Rbuildignore} is updated to ignore this file.
\item Adds recommended settings to \code{.vscode/settings.json} and
\code{.vscode/extensions.json}. This includes:
\itemize{
\item Setting \code{editor.formatOnSave = true} for R files to enable formatting on
every save. Does nothing if \code{editor.formatOnSave} was already set.
\item Setting \code{editor.defaultFormatter} to Air for R files to ensure that Air
is always selected as the formatter for this project. Does nothing if
\code{editor.defaultFormatter} was already set.
\item Setting the Air extension as a "recommended" extension for this project,
which triggers a notification for contributors coming to this project
that don't yet have the Air extension installed.
}

If \verb{.vscode/} does not exist yet, it is created. If the project is an R
package, \code{.Rbuildignore} is updated to ignore this folder.

If you'd like to opt out of VS Code / Positron specific setup, set \code{vscode = FALSE}, but remember that even if you work in RStudio, other contributors
may prefer another editor.
}

Note that this setup does not install an Air binary, so there is an
additional manual step you must take before using Air for the first time:
\itemize{
\item For RStudio, follow the \href{https://posit-dev.github.io/air/editor-rstudio.html}{installation guide}.
\item For Positron, install the \href{https://open-vsx.org/extension/posit/air-vscode}{OpenVSX Extension}.
\item For VS Code, install the \href{https://marketplace.visualstudio.com/items?itemName=Posit.air-vscode}{VS Code Extension}.
\item For other editors, check to \href{https://posit-dev.github.io/air/editors.html}{see if that editor is supported} by Air.
}
}
\examples{
\dontrun{
# Prepare an R package or project to use Air
use_air()
}
}
