% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoconf.R
\name{use_autoconf}
\alias{use_autoconf}
\title{Use Autoconf}
\usage{
use_autoconf(open = interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Creates an example \code{configure.ac} file that can be used to configure how a
package is built by R during installation. This is primarily useful when you
are writing a package that uses some system C or C++ libraries, and want to
detect the presence of these libraries during installation.
}
\details{
This function also sets up various other \code{autoconf}-releated files, namely by
generating a \code{cleanup} script and a \code{src/Makevars.in} file. It will also
create a git commit hook (if git appears to be used by the package) and add
various files to the ignore files.

You may want to consult R's \href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Configure-example}{documentation}
on configure scripts for additional examples.
}
\examples{
\dontrun{
use_autoconf()
}
}
