% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{use_badge}
\alias{use_badge}
\title{Use a README badge}
\usage{
use_badge(badge_name, href, src, base_path = ".")
}
\arguments{
\item{badge_name}{Badge name. Used in error message and alt text}

\item{href, src}{Badge link and image src}

\item{base_path}{Path to package root.}
}
\description{
Use a README badge
}
