% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{use_binder_badge}
\alias{use_binder_badge}
\title{Create a Binder badge}
\usage{
use_binder_badge()
}
\description{
This prints out the markdown which will display a Binder "badge", indicating that
your repository can be launched in an executable environment on \url{https://mybinder.org/}.
}
