% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff.R
\name{use_cff}
\alias{use_cff}
\title{Create Citation File Format}
\usage{
use_cff()
}
\description{
\code{use_cff()} ...
}
\examples{
\dontrun{
use_cff()
}
}
\seealso{
The \href{https://r-pkgs.org/data.html}{data chapter} of \href{https://r-pkgs.org}{R Packages}.
}
