% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clang-format.R
\name{use_clang_format}
\alias{use_clang_format}
\title{Use clang-format}
\usage{
use_clang_format(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\description{
Creates \code{src/.clang-format} and adds a before_install step to
`.travis.yml` to check the format is followed.
}
