% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-of-conduct.R
\name{use_code_of_conduct}
\alias{use_code_of_conduct}
\title{Add a code of conduct}
\usage{
use_code_of_conduct()
}
\description{
The goal of a code of conduct is to foster an environment of inclusiveness,
and to explicit discourage inappropriate behaviour. This code of conduct
comes from \url{http://contributor-covenant.org}, version 1:
\url{http://contributor-covenant.org/version/1/0/0/}.
}
