% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-compat.R
\name{use_compat}
\alias{use_compat}
\title{Use a compatibility file from another repo}
\usage{
use_compat(repo_spec, file = NULL)
}
\arguments{
\item{repo_spec}{A string identifying the GitHub repo in one of these forms:
\itemize{
\item Plain \code{OWNER/REPO} spec
\item Browser URL, such as \code{"https://github.com/OWNER/REPO"}
\item HTTPS Git URL, such as \code{"https://github.com/OWNER/REPO.git"}
\item SSH Git URL, such as \code{"git@github.com:OWNER/REPO.git"}
}

In the case of a browser, HTTPS, or SSH URL, the \code{host} is extracted from
the URL. The \code{REPO} part will be the name of the new local folder, which is
also a project and Git repo.}

\item{file}{Name of compat file. The \verb{compat-} suffix and file extension
are optional. If ommitted, will allow you to choose from the compat
files offered by that repo.}
}
\description{
A "compat" file implements a minimum set of standalone functionality in
such a way that it can be copied into another package. \code{use_compat()}
makes it easy to get such a file into your own repo.
}
