% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{use_course}
\alias{use_course}
\title{Download course materials}
\usage{
use_course(url, destdir = NULL)
}
\arguments{
\item{url}{Link to a ZIP file containing the materials, possibly behind a
shortlink. Function developed with DropBox and GitHub in mind, but should
work for ZIP files generally. See \code{\link[=download_zip]{download_zip()}} for more.}

\item{destdir}{The new folder is stored here. Defaults to working directory.}
}
\value{
Path to the new directory holding the course materials.
}
\description{
Special-purpose function to download a folder of course materials. The only
demand on the user is to confirm or specify where the new folder should be
stored. Workflow:
\itemize{
\item User calls \code{use_course("SHORTLINK-GOES-HERE")}.
\item User is asked to notice and confirm the location of the new folder Specify
\code{destdir} to skip this.
\item User is asked if they'd like to delete the ZIP file.
\item New folder is opened in the file manager, e.g. Finder or File Explorer.
}
}
\examples{
\dontrun{
## demo with a small CRAN package available in various places

## from CRAN
use_course("https://cran.r-project.org/bin/windows/contrib/3.4/rematch2_2.0.1.zip")

## from GitHub, 2 ways
use_course("https://github.com/r-lib/rematch2/archive/master.zip")
use_course("https://api.github.com/repos/r-lib/rematch2/zipball/master")
}
}
\seealso{
Other download functions: \code{\link{download_zip}},
  \code{\link{tidy_unzip}}
}
\concept{download functions}
