% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11.R
\name{use_cpp11}
\alias{use_cpp11}
\title{Use C++ via the cpp11 package}
\usage{
use_cpp11(name = NULL)
}
\arguments{
\item{name}{If supplied, creates and opens \verb{src/name.\{c,cpp\}}.}
}
\description{
Adds infrastructure commonly needed when using compiled code:
\itemize{
\item Creates \verb{src/}
\item Adds required packages to \code{DESCRIPTION}
\item May create an initial placeholder \code{.c} or \code{.cpp} file
}
}
\details{
When using compiled code, please note that there must be at least one file
inside the \verb{src/} directory prior to building the package. As a result,
if an empty \verb{src/} directory is detected, either a \code{.c} or \code{.cpp} file will
be added.
}
