% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{use_cran_badge}
\alias{use_cran_badge}
\title{Create a CRAN badge}
\usage{
use_cran_badge()
}
\description{
This prints out the markdown which will display a CRAN "badge", indicating
what version of your package is available on CRAN, powered by
\url{http://www.r-pkg.org}.
}
