% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pronoun.R
\name{use_data_pronoun}
\alias{use_data_pronoun}
\title{Use the tidy eval data pronoun in your package}
\usage{
use_data_pronoun()
}
\description{
Does setup necessary to use rlang's \code{.data} pronoun in your package. See \link[rlang:.data]{rlang::.data} for more.
This function requires the use roxygen.
\itemize{
\item Adds rlang to "Imports" in \code{DESCRIPTION}.
\item Imports the \code{.data} object specifically, which is necessary for internal
use.
}
}
\examples{
\dontrun{
use_data_pronoun()
}
}
