% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{use_depsy_badge}
\alias{use_depsy_badge}
\title{Create a Depsy badge}
\usage{
use_depsy_badge(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\description{
This prints out the markdown which will display a Depsy "badge", showing the
"percentile overall impact" of the project, powered by
\url{http://www.depsy.org}.
}
\details{
Depsy only indexes projects that are on CRAN.
}
