% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{use_example}
\alias{use_example}
\title{Create or edit example files}
\usage{
use_example(name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{name}{Either a string giving a file name (without directory) or
\code{NULL} to take the name from the currently open file in RStudio.}

\item{open}{Whether to open the file for interactive editing.}
}
\description{
This function makes it easy to create paired R and example files,
using the convention that the examples for \code{R/foofy.R} should live
in \code{man/examples/example-foofy.R}. You can use it to create a new file
from scratch by supplying \code{name}, or if you use RStudio, you can call
to create (or navigate to) the example file based on the currently open
script.
}
\section{Including an example file in roxygen documentation}{


To add examples stored in an example file to roxygen documentation,
add the following line:

\if{html}{\out{<div class="sourceCode default">}}\preformatted{#' @example man/examples/example-<file>.R
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link[=use_r]{use_r()}} and \code{\link[=use_test]{use_test()}} to create R and test files
using a similar pattern.
}
