% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-branch-default.R
\name{use_git_branch_default}
\alias{use_git_branch_default}
\title{Change (or discover a change in) the default Git branch}
\usage{
use_git_branch_default(
  scope = c("project", "user"),
  old_name = NULL,
  new_name = NULL
)
}
\arguments{
\item{scope}{Edit globally for the current \strong{user}, or locally for the
current \strong{project}}

\item{old_name}{Old name of the default branch. If unspecified, this can
often be inferred.}

\item{new_name}{New name of the default branch. The presence of this argument
determines whether we are actively changing the project's default branch or
discovering whether such a change has been made in the project.}
}
\value{
The new default branch name.
}
\description{
While Git itself has no official concept of "default branch", in reality,
practically all Git repos and users have an \strong{effective default branch}.
Historically this was usually \code{master}, but \code{main} is an increasingly popular
choice and this branch renaming requires some local and remote work.

\code{use_git_branch_default()} facilitates several proactive steps around the
default branch:
\itemize{
\item \code{use_git_branch_default(scope = "user", new_name = "main")} configures
\code{init.defaultBranch} at the global (a.k.a user) level. This only affects
new local Git repos you create in the future.
\item \code{use_git_branch_default(scope = "project")} discovers the default branch
from the remote source repo for the current Git repo / project. If that
appears to have changed, e.g. from \code{master} to \code{main}, we do the necessary
branch renaming in your local repo and, if relevant, in your fork.
\item \code{use_git_branch_default(scope = "project", old_name = "master", new_name = "main")}
moves the local \code{master} branch to \code{main}, updates the remote source repo,
and, if relevant, also updates your fork. This only works for a repo that
you administer.
}
}
\examples{
\dontrun{
use_git_branch_default(scope = "user", new_name = "main")

use_git_branch_default(scope = "project")

use_git_branch_default(scope = "project", old_name = "master", new_name = "main")
}
}
