% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{use_git_ignore}
\alias{use_git_ignore}
\title{Tell git to ignore files}
\usage{
use_git_ignore(ignores, directory = ".", base_path = ".")
}
\arguments{
\item{directory}{Directory within \code{base_path} to set ignores}

\item{base_path}{Path to package root.}

\item{ignore}{Character vector of ignores, specified as file globs.}
}
\description{
Tell git to ignore files
}
\seealso{
Other git.helpers: \code{\link{use_git_hook}},
  \code{\link{use_git}}
}

