% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{use_git_protocol}
\alias{use_git_protocol}
\title{Choose default git protocol and credential defaults}
\usage{
use_git_protocol(protocol = NULL)
}
\arguments{
\item{protocol}{"ssh" or "https". Defaults to the option \code{usethis.protocol}
and, if unset, to an interactive choice or, in non-interactive sessions,
"ssh".}
}
\value{
"ssh" or "https"
}
\description{
Decide whether to use the SSH or HTTPS protocol for the current session.
This affects two things:
\itemize{
\item The URL format of newly configured git remotes:
\itemize{
\item \code{protocol = "ssh"} implies \code{git@github.com:<OWNER>/<REPO>.git}
\item \code{protocol = "https"} implies \code{https://github.com/<OWNER>/<REPO>.git}
}
\item The strategy for creating default credentials when none are given.
Credentials are needed for git operations like \code{git push} that address a
remote, typically GitHub. Remember you can always pass explicit
\code{credentials} made with \code{\link[git2r:cred_ssh_key]{git2r::cred_ssh_key()}}, \link[git2r:cred_user_pass]{git2r::cred_user_pass},
and friends.
This function also gives a reminder of how to set a default protocol via an
option: \code{options(usethis.protocol = "ssh")} or
\code{options(usethis.protocol = "https")}.
}
}
\section{Default credentials}{


The current git protocol is consulted to create default credentials when
they are needed but left unspecified.

For \code{protocol = "ssh"}, usethis passes \code{NULL} credentials straight through to
git2r. This will do the right thing if you have the exact configuration
expected by git2r: your public and private keys are in the default locations,
\code{~/.ssh/id_rsa.pub} and \code{~/.ssh/id_rsa}, respectively, and your \code{ssh-agent}
is configured to manage any associated passphrase, if relevant. Read more
about SSH setup in \href{http://happygitwithr.com/ssh-keys.html}{Happy Git},
especially the \href{http://happygitwithr.com/ssh-keys.html#ssh-troubleshooting}{troubleshooting section}.

For \code{protocol = "https"}, usethis attempts to use a GitHub personal access
token (PAT), with preference for an \code{auth_token} that is passed explicitly.
If that is \code{NULL}, the environment variables \code{GITHUB_PAT} and \code{GITHUB_TOKEN}
are consulted, in that order. If a PAT is found, we make an HTTPS credential
with \code{\link[git2r:cred_user_pass]{git2r::cred_user_pass()}}. The PAT is sent as the password and dummy
text is sent as the username (only the PAT is truly required).
}

\examples{
\dontrun{
use_git_protocol()
use_git_protocol("ssh")
use_git_protocol("https")
}
}
