% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{use_git_remote}
\alias{use_git_remote}
\alias{git_remotes}
\title{Configure and report Git remotes}
\usage{
use_git_remote(name = "origin", url, overwrite = FALSE)

git_remotes()
}
\arguments{
\item{name}{A string giving the short name of a remote.}

\item{url}{A string giving the url of a remote.}

\item{overwrite}{Logical. Controls whether an existing remote can be
modified.}
}
\value{
Named list of Git remotes.
}
\description{
Two helpers are available:
\itemize{
\item \code{use_git_remote()} sets the remote associated with \code{name} to \code{url}.
\item \code{git_remotes()} reports the configured remotes, similar to
\code{git remote -v}.
}
}
\examples{
\dontrun{
## see current remotes
git_remotes()

## add a new remote, a la `git remote add <name> <url>`
use_git_remote(name = "origin", url = "https://github.com/<OWNER>/<REPO>.git")

## remove an existing remote, a la `git remote remove <name>`
use_git_remote(name = "foo", url = NULL, overwrite = TRUE)

## change URL of existing remote, a la `git remote set-url <name> <newurl>`
use_git_remote(
  name = "foo",
  url = "https://github.com/<OWNER>/<REPO>.git",
  overwrite = TRUE
)
}
}
