% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{use_github_actions}
\alias{use_github_actions}
\alias{use_github_action_check_release}
\alias{use_github_action_check_standard}
\alias{use_github_action_pr_commands}
\alias{use_github_action_check_full}
\title{Deprecated GitHub Actions functions}
\usage{
use_github_actions()

use_github_action_check_release(
  save_as = "R-CMD-check.yaml",
  ref = NULL,
  ignore = TRUE,
  open = FALSE
)

use_github_action_check_standard(
  save_as = "R-CMD-check.yaml",
  ref = NULL,
  ignore = TRUE,
  open = FALSE
)

use_github_action_pr_commands(
  save_as = "pr-commands.yaml",
  ref = NULL,
  ignore = TRUE,
  open = FALSE
)

use_github_action_check_full(
  save_as = "R-CMD-check.yaml",
  ignore = TRUE,
  open = FALSE,
  repo_spec = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item \code{use_github_actions()} is deprecated because it was just an alias
for \code{\link[=use_github_action_check_release]{use_github_action_check_release()}}.
\item \code{use_github_action_check_full()} is overkill for most packages and is
not recommended.
\item \code{use_github_action_check_release()}, \code{use_github_action_check_standard()},
and \code{use_github_action_pr_commands()} are deprecated in favor of
\code{\link[=use_github_action]{use_github_action()}}, which can now suggest specific workflows to use.
}
}
\keyword{internal}
