% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-actions.R
\name{use_github_actions_badge}
\alias{use_github_actions_badge}
\title{Generates a GitHub Actions badge}
\usage{
use_github_actions_badge(name = "R-CMD-check.yaml", repo_spec = NULL)
}
\arguments{
\item{name}{Name of the workflow's YAML configuration file (with or without
extension), e.g. \code{"R-CMD-check"}, \code{"R-CMD-check.yaml"}.}
}
\description{
Generates a GitHub Actions badge and that's all. This exists primarily for
internal use.
}
\keyword{internal}
