% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-labels.R
\name{use_github_labels}
\alias{use_github_labels}
\alias{gh_labels}
\title{Manage GitHub issue labels.}
\usage{
use_github_labels(labels = gh_labels(), delete_default = FALSE,
  auth_token = NULL, host = NULL)

gh_labels()
}
\arguments{
\item{labels}{Named character vector of labels. The names are the label text,
such as "bug", and the values are the label colours in hexadecimal, such as
"d73a4a". First, labels that don't yet exist are created, then label
colours are updated.}

\item{delete_default}{If \code{TRUE}, will remove GitHub default labels that do
not appear in the \code{labels} vector (presumably defaults that aren't relevant
to your workflow).}

\item{auth_token}{Provide a personal access token (PAT) from
\url{https://github.com/settings/tokens}. If \code{NULL}, will use the
\code{GITHUB_PAT} environment variable.}

\item{host}{GitHub API host to use. Override with the endpoint-root for your
GitHub enterprise instance, for example,
"https://github.hostname.com/api/v3". You can set this globally using
the \code{GITHUB_API_URL} env var.}
}
\description{
\code{use_github_labels()} creates new labels and/or changes label
colours. It does not generally remove labels, unless you explicitly ask to
remove the default GitHub labels that are not present in the labels you
provide via \code{labels}.

\code{gh_labels()} returns the labels and colours commonly used by tidyverse
packages.
}
\examples{
\dontrun{
## typical use in, e.g., a new tidyverse project
use_github_labels(delete_default = TRUE)
}
}
