% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-pages.R
\name{use_github_pages}
\alias{use_github_pages}
\title{Configure a GitHub Pages site}
\usage{
use_github_pages(branch = "gh-pages", path = "/", cname = NA)
}
\arguments{
\item{branch}{Branch for the site source. Default is \code{gh-pages}, coupled with
\code{path = "/"}. These are the usethis defaults, but more importantly, they
are GitHub's defaults, i.e. when a \code{gh-pages} branch is first created, it
is \emph{automatically} published to Pages, using the source found in \code{"/"}. If
a \code{gh-pages} branch does not yet exist on the host, \code{use_github_pages()}
creates an empty, orphan remote branch.

The most common alternative is the repo's default branch, coupled with
\code{path = "/docs"}. It is the user's responsibility to ensure that this
\code{branch} pre-exists on the host.}

\item{path}{Path to the site source. Note that GitHub only supports \code{"/"} or
\verb{"/docs".}}

\item{cname}{Optional, custom domain name. The \code{NA} default means "don't set
or change this", whereas a value of \code{NULL} removes any previously
configured custom domain.}
}
\value{
Nothing
}
\description{
Activates or reconfigures a GitHub Pages site for a project hosted on GitHub.
This function anticipates two specific usage modes:
\itemize{
\item Publish from the root directory of a \code{gh-pages} branch, which is assumed to
be only (or at least primarily) a remote branch.
\itemize{
\item Typically the \code{gh-pages} branch is managed by an automatic "build and
deploy" job, such as the one configured by
\code{use_github_action("pkgdown")}.
}
\item Publish from the \code{"/docs"} directory of a "regular" branch, probably the
repo's default branch.
\itemize{
\item The user is assumed to have a plan for how they will manage the content
below \code{"/docs"}.
}
}
}
\examples{
\dontrun{
use_github_pages()
}
}
\seealso{
\itemize{
\item \url{https://docs.github.com/en/free-pro-team@latest/github/working-with-github-pages}
\item \url{https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#pages}
}
}
