% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenkinsfile.R
\name{use_jenkinsfile}
\alias{use_jenkinsfile}
\title{Create Jenkinsfile for Jenkins CI Pipelines}
\usage{
use_jenkinsfile()
}
\description{
\code{use_jenkinsfile()} adds a basic Jenkinsfile for R packages to the project
root directory. The Jenkinsfile stages take advantage of calls to \code{make}, and
so calling this function will also run \code{use_makefile()} if a Makefile does
not already exist at the project root.
}
\seealso{
The \href{https://jenkins.io/doc/book/pipeline/jenkinsfile/}{documentation on Jekins Pipelines}.

\code{\link[=use_makefile]{use_makefile()}}
}
