% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{use_lifecycle_badge}
\alias{use_lifecycle_badge}
\title{Create a life cycle badge}
\usage{
use_lifecycle_badge(stage)
}
\arguments{
\item{stage}{Stage of the lifecycle. See description above.}
}
\description{
Declares the developmental stage of a package, according to
\url{https://www.tidyverse.org/lifecycle/}:
}
\details{
\itemize{
\item Experimental
\item Maturing
\item Stable
\item Retired
\item Archived
\item Dormant
\item Questioning
}
}
