% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linting.R
\name{use_lintr}
\alias{use_lintr}
\title{Use code linting}
\usage{
use_lintr(type = c("local", "travis", "test"))
}
\arguments{
\item{type}{Options are \code{local}, \code{travis} or \code{test}.}
}
\description{
\code{use_lintr()} Adds support for code linting using the \code{lintr} package in one of three ways:
local interactive linting only, non-failing (post-build) Travis CI linting, or failing local and Travis CI \code{testthat} unit test.
}
