% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{use_logo}
\alias{use_logo}
\title{Use a package logo}
\usage{
use_logo(img, geometry = "240x278", retina = TRUE)
}
\arguments{
\item{img}{The path to an existing image file}

\item{geometry}{a \link[magick:geometry]{magick::geometry} string specifying size. The default
assumes that you have a hex logo using spec from
\verb{http://hexb.in/sticker.html}.}

\item{retina}{\code{TRUE}, the default, scales the image on the README,
assuming that geometry is double the desired size.}
}
\description{
This function helps you use a logo in your package:
\itemize{
\item Enforces a specific size
\item Stores logo image file at \code{man/figures/logo.png}
\item Produces the markdown text you need in README to include the logo
}
}
\examples{
\dontrun{
use_logo("usethis.png")
}
}
