% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{use_make}
\alias{use_make}
\title{Create Makefile}
\usage{
use_make()
}
\description{
\code{use_make()} adds a basic Makefile to the project root directory.
}
\seealso{
The \href{https://www.gnu.org/software/make/manual/html_node/}{documentation for GNU Make}.
}
