% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile.R
\name{use_makefile}
\alias{use_makefile}
\title{Create Makefile}
\usage{
use_makefile()
}
\description{
\code{use_makefile()} adds a basic Makefile to the project root directory.
}
\seealso{
The \href{https://www.gnu.org/software/make/manual/html_node}{documentation for GNU Make}.
}
