% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{use_namespace}
\alias{use_namespace}
\title{Use a basic \code{NAMESPACE}}
\usage{
use_namespace(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\description{
This \code{NAMESPACE} exports everything, except functions that start
with a \code{.}.
}

