% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{use_pipe}
\alias{use_pipe}
\title{Use magrittr's pipe in your package}
\usage{
use_pipe(export = TRUE)
}
\arguments{
\item{export}{If \code{TRUE}, the file \code{R/utils-pipe.R} is added, which provides
the roxygen template to import and re-export \verb{\%>\%}. If \code{FALSE}, the necessary
roxygen directive is added, if possible, or otherwise instructions are given.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The base R pipe has been available since R 4.1.0 and we recommend using it
in all actively maintained and new work, both inside and outside packages.
The \code{magrittr} pipe (\verb{\%>\%}) can usually just be replaced with base R's \verb{|>},
with a few things to keep in mind:
\itemize{
\item Instead of \code{x \%>\% f}, use \code{x |> f()}.
\item Instead of \code{x \%>\% f(1, y = .)}, use \code{x |> f(1, y = _)}.
\item Instead of \code{x \%>\% f(a = ., b = .)}, define a new helper function.
}

Learn more in \url{https://tidyverse.org/blog/2023/04/base-vs-magrittr-pipe/}.
}
\examples{
\dontrun{
use_pipe()
}
}
\keyword{internal}
