% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{use_ragg}
\alias{use_ragg}
\title{Use AGG as RStudio's Graphic Backend}
\usage{
use_ragg()
}
\description{
\code{use_ragg()} sets the user's default graphic device in RStudio to use AGG
(via the ragg R package). Compared to the default graphics device, AGG is
faster, higher quality, and more consistent across operating systems. More
detail can be found in the blog post \href{https://www.tidyverse.org/blog/2021/02/modern-text-features/}{Modern Text Features in R}. Setting the
backend graphics device only affects plots generated by RStudio; to use ragg
in R Markdown, set \code{dev = "ragg_png"} in your chunk options.
}
