% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp.R
\name{use_rcpp}
\alias{use_rcpp}
\alias{use_rcpp_armadillo}
\alias{use_rcpp_eigen}
\alias{use_c}
\title{Use C, C++, RcppArmadillo, or RcppEigen}
\usage{
use_rcpp(name = NULL)

use_rcpp_armadillo(name = NULL)

use_rcpp_eigen(name = NULL)

use_c(name = NULL)
}
\arguments{
\item{name}{Either a string giving a file name (without directory) or
\code{NULL} to take the name from the currently open file in RStudio or
Positron.}
}
\description{
Adds infrastructure commonly needed when using compiled code:
\itemize{
\item Creates \verb{src/}
\item Adds required packages to \code{DESCRIPTION}
\item May create an initial placeholder \code{.c} or \code{.cpp} file
\item Creates \code{Makevars} and \code{Makevars.win} files (\code{use_rcpp_armadillo()} only)
}
}
