% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_armadillo.R
\name{use_rcpp_armadillo}
\alias{use_rcpp_armadillo}
\title{Use RcppArmadillo}
\usage{
use_rcpp_armadillo()
}
\description{
Creates \code{src/}, adds needed packages to \code{DESCRIPTION},
and configures \code{src/Makevars} and \code{src/Makevars.win}.
}
