% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{use_release_issue}
\alias{use_release_issue}
\title{Create a release checklist in a GitHub issue}
\usage{
use_release_issue(version = NULL)
}
\arguments{
\item{version}{Optional version number for release. If unspecified, you can
make an interactive choice.}
}
\description{
When preparing to release a package to CRAN there are quite a few steps that
need to be performed, and some of the steps can take multiple hours. This
function creates a checklist in a GitHub issue to:
\itemize{
\item Help you keep track of where you are in the process
\item Feel a sense of satisfaction as you progress towards final submission
\item Help watchers of your package stay informed.
}

The checklist contains a generic set of steps that we've found to be helpful,
based on the type of release ("patch", "minor", or "major"). You're
encouraged to edit the issue to customize this list to meet your needs.
\subsection{Customization}{
\itemize{
\item If you want to consistently add extra bullets for every release, you can
include your own custom bullets by providing an (unexported)
\code{release_bullets()} function that returns a character vector.
(For historical reasons, \code{release_questions()} is also supported).
\item If you want to check additional packages in the revdep check process,
provide an (unexported) \code{release_extra_revdeps()} function that
returns a character vector. This is currently only supported for
Posit internal check tooling.
}
}

\subsection{Built-in bullets}{

Prepare for release:
\itemize{
\item \verb{git pull}
\item Check \href{https://cran.rstudio.org/web/checks/check_results_usethis.html}{current CRAN check results}
\item \href{https://lifecycle.r-lib.org/articles/communicate.html#gradual-deprecation}{Advance deprecations}, if needed
\item \href{https://style.tidyverse.org/news.html#news-release}{Polish NEWS}
\item \code{urlchecker::url_check()}
\item \code{devtools::build_readme()}
\item \code{devtools::check(remote = TRUE, manual = TRUE)}
\item \code{devtools::check_win_devel()}
\item \code{revdepcheck::cloud_check()}
\item Update \code{cran-comments.md}
\item \verb{git push}
\item Draft blog post
\item Check that \code{use_code_of_conduct()} is shipping the latest version of the Contributor Covenant (\url{https://www.contributor-covenant.org}).
}

Submit to CRAN:
\itemize{
\item \code{usethis::use_version('major')}
\item \code{devtools::submit_cran()}
\item Approve email
}

Wait for CRAN...
\itemize{
\item Accepted :tada:
\item Finish & publish blog post
\item Add link to blog post in pkgdown news menu
\item \code{usethis::use_github_release()}
\item \code{usethis::use_dev_version(push = TRUE)}
\item Share on social media
\item Slack link to blog post, bluesky, and linkedin in #open-source-comms
}
}
}
\examples{
\dontrun{
use_release_issue("2.0.0")
}
}
