% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_status.R
\name{use_repo_status_badge}
\alias{use_repo_status_badge}
\title{Create a Repository badge}
\usage{
use_repo_status_badge(status = c("active", "abandoned", "concept", "inactive",
  "moved", "suspended", "unsupported", "wip"))
}
\arguments{
\item{status}{Status of repository, see \url{http://www.repostatus.org}}
}
\description{
This prints out the markdown which will display a Repo "badge", indicating
the states of usability and development, powered by
\url{http://www.repostatus.org}.
}
