% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep.R
\name{use_revdep}
\alias{use_revdep}
\title{Set up devtools revdep template}
\usage{
use_revdep(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\description{
Add \code{revdep} directory and basic check template.
}

