% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{use_rmarkdown_template}
\alias{use_rmarkdown_template}
\title{Add an RMarkdown Template}
\usage{
use_rmarkdown_template(template_dir = "newtemplate",
  template_name = "Template Name",
  template_description = "A description of the template",
  template_create_dir = FALSE, base_path = ".")
}
\arguments{
\item{template_dir}{Name of the directory the template will live in within
\code{inst/rmarkdown/templates}.}

\item{template_name}{The name as printed in the template menu.}

\item{template_description}{A short description of the template.}

\item{template_create_dir}{Sets the value of \code{create_dir}
in the template.yml}

\item{base_path}{Path to package root.}
}
\description{
Add an RMarkdown Template
}
\examples{
\dontrun{
use_rmarkdown_template()
}
}
