% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-deprecated.R
\name{use_rscloud_badge}
\alias{use_rscloud_badge}
\title{Deprecated badge function}
\usage{
use_rscloud_badge(url)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{use_rscloud_badge()} has been replaced by \code{\link[=use_posit_cloud_badge]{use_posit_cloud_badge()}}.
}
\keyword{internal}
