% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{use_rstudio}
\alias{use_rstudio}
\title{Add RStudio Project infrastructure}
\usage{
use_rstudio()
}
\description{
This function makes the current project into an RStudio Project:
\itemize{
\item Creates an \code{.Rproj} file
\item Adds RStudio files to \code{.gitignore}
\item Adds RStudio files to \code{.Rbuildignore}, if project is a package
}
}
