% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{use_rstudio_preferences}
\alias{use_rstudio_preferences}
\title{Set global RStudio preferences}
\usage{
use_rstudio_preferences(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Property-value pairs.}
}
\value{
A named list of the previous values, invisibly.
}
\description{
This function allows you to set global RStudio preferences, achieving the
same effect programmatically as clicking buttons in RStudio's Global Options.
You can find a list of configurable properties at
\url{https://docs.posit.co/ide/server-pro/reference/session_user_settings.html}.
}
