% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7.R
\name{use_s7}
\alias{use_s7}
\title{Use S7}
\usage{
use_s7(backwards_compat = TRUE)
}
\arguments{
\item{backwards_compat}{If \code{TRUE} (the default), adds a \verb{@rawNamespace}
directive to the package-level documentation that conditionally imports
the \code{@} operator from S7 for R versions prior to 4.3.0.}
}
\description{
Sets up a package to use \href{https://rconsortium.github.io/S7/}{S7} classes.
\itemize{
\item Adds S7 to \code{Imports} in \code{DESCRIPTION}
\item Creates \code{R/zzz.R} with a call to \code{S7::methods_register()} in \code{.onLoad()}
\item Optionally adds a \verb{@rawNamespace} directive to enable the use of
\verb{<S7_object>@name} syntax in package code for R versions prior to 4.3.0
(see \href{https://rconsortium.github.io/S7/articles/packages.html}{Using S7 in a Package})
}
}
\examples{
\dontrun{
use_s7()
}
}
