% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stale.R
\name{use_stale}
\alias{use_stale}
\title{Add Stale Bot}
\usage{
use_stale(open = interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
\code{use_stale()} Adds \href{https://github.com/apps/stale}{GitHub Stale Bot} to
add Stale bot to your repository
}
