% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{use_test_helper}
\alias{use_test_helper}
\title{Create or edit a test helper file}
\usage{
use_test_helper(open = rlang::is_interactive())
}
\arguments{
\item{open}{Whether to open the file for interactive editing.}
}
\description{
This function helps create a test helper file, namely a file named
\code{tests/testthat/helper.R}. Test helper files are functions loaded with
\code{devtools::load_all()} that are  helpful to run tests.
}
\seealso{
\itemize{
\item \code{\link[=use_test]{use_test()}} to create tests.
\item The testthat vignette on special files
\code{vignette("special-files", package = "testthat")}.
}
}
