% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{use_testthat}
\alias{use_testthat}
\title{Sets up overall testing infrastructure
creates
\code{tests/testthat/}, \code{tests/testthat.R}, and adds testthat to Suggests.}
\usage{
use_testthat()
}
\description{
Sets up overall testing infrastructure
creates
\code{tests/testthat/}, \code{tests/testthat.R}, and adds testthat to Suggests.
}
\examples{
\dontrun{
use_testthat()

use_test()

use_test("something-management")
}
}
\seealso{
\link{use_test} to create test files. The \href{https://r-pkgs.org/tests.html}{testing chapter} of \href{https://r-pkgs.org}{R Packages}.
}
