% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{use_tibble}
\alias{use_tibble}
\title{Use tidyverse tibbles in your package}
\usage{
use_tibble()
}
\description{
Does setup necessary to use the tidyverse's tibble internally in your package and to
enable tibble printing out of the box:
\itemize{
\item Adds tibble to "Imports" in DESCRIPTION
\item Creates \code{R/utils-tibble.R} with the necessary roxygen template
}
}
