% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.R
\name{use_tic}
\alias{use_tic}
\title{Set up CI with the tic package}
\usage{
use_tic(services = NULL)
}
\arguments{
\item{services}{\code{[character]}\cr
CI services to add, default: auto-detect.}
}
\description{
By default the CI-services "Travis" (Linux) and "Appveyor"
(Windows) will be set up. Basic \code{.travis.yml} and \code{appveyor.yml} files are
added to the top-level directory of a package.
}
\details{
This function is aimed at supporting the most common use cases.
Users who require more control are advised to manually call the individual
functions.

The following steps will be run:
\enumerate{
\item If necessary, create a GitHub repository via \code{\link[=use_github]{use_github()}}
\item Enable Travis via \code{\link[travis:travis_enable]{travis::travis_enable()}}
\item Create a default \code{.travis.yml} file
(overwrite after confirmation in interactive mode only)
\item Create a default \code{appveyor.yml} file
(if requested, by default only for packages, overwrite after confirmation
in interactive mode only)
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
\item Enable deployment (if necessary, depending on repo type)
via \code{\link[=use_travis_deploy]{use_travis_deploy()}}
\item Create a GitHub PAT and install it on Travis CI via \code{\link[travis:travis_set_pat]{travis::travis_set_pat()}}
}
}
