% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-agents.R
\name{use_tidy_agents}
\alias{use_tidy_agents}
\title{Configure a project to work with AI coding agents}
\usage{
use_tidy_agents()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function sets up a project to work with AI coding agents.
Specifically, it:
\itemize{
\item Creates an \code{AGENTS.md} file containing project-specific instructions for
R package development. It includes general development advice, as well
as pointers to specific skills provided by \code{\link[=learn_tidy_skill]{learn_tidy_skill()}}.

This file is read by most coding agents, including Posit Assistant,
Codex, Gemini CLI, and Cursor.

The file begins with a "This package" section for your own
package-specific advice. If you re-run \code{use_tidy_agents()} to update
\code{AGENTS.md}, the contents of this section are preserved.
\item \code{.Rbuildignore} ignores \code{AGENTS.md}, so it isn't included in your built
package.
}
}
\examples{
\dontrun{
use_tidy_agents()
}
}
