% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-deprecated.R
\name{use_tidy_eval}
\alias{use_tidy_eval}
\title{Deprecated tidyverse functions}
\usage{
use_tidy_eval()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item \code{use_tidy_eval()} is deprecated because there's no longer a need to
systematically import and re-export a large number of functions in order
to use tidy evaluation. Instead, use \code{\link[=use_import_from]{use_import_from()}} to tactically
import functions as you need them.
}
}
\keyword{internal}
