% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{use_tidy_labels}
\alias{use_tidy_labels}
\alias{use_tidy_ci}
\alias{use_tidy_eval}
\title{Defunct tidyverse functions}
\usage{
use_tidy_labels()

use_tidy_ci(...)

use_tidy_eval()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
\itemize{
\item \code{use_tidy_labels()} has been replaced by \code{\link[=use_tidy_github_labels]{use_tidy_github_labels()}}.
\item \code{use_tidy_ci()} has been replaced by \code{\link[=use_tidy_github_actions]{use_tidy_github_actions()}}.
\item \code{use_tidy_eval()} is defunct because there's no longer a need to
systematically import and re-export a large number of functions in order
to use tidy evaluation. Instead, use \code{\link[=use_import_from]{use_import_from()}} to tactically
import functions as you need them.
}
}
\keyword{internal}
