% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{use_tidy_style}
\alias{use_tidy_style}
\title{Style according to the tidyverse style guide}
\usage{
use_tidy_style(strict = TRUE)
}
\arguments{
\item{strict}{Boolean indicating whether or not a strict version of styling
should be applied. See \code{\link[styler:tidyverse_style]{styler::tidyverse_style()}} for details.}
}
\value{
Invisibly returns a data frame that indicates for each file considered for
styling whether or not it was actually changed.
}
\description{
Styles the source code of the package according to the
\href{http://style.tidyverse.org}{tidyverse style guide}.
}
\section{Warning}{

This function overwrites files (if styling results in a change of the
code to be formatted). It is strongly suggested to only style files
that are under version control or to create a backup copy.
}

