% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{use_travis}
\alias{use_travis}
\alias{use_travis_badge}
\alias{use_pkgdown_travis}
\alias{use_appveyor}
\alias{use_appveyor_badge}
\title{CI on Travis and Appveyor}
\usage{
use_travis(browse = rlang::is_interactive(), ext = c("com", "org"))

use_travis_badge(ext = c("com", "org"), repo_spec = NULL)

use_pkgdown_travis()

use_appveyor(browse = rlang::is_interactive())

use_appveyor_badge(repo_spec = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions which formally supported CI on Appveyor and Travis are
now defunct as we no longer recommend using these services. We now
recommend using GitHub actions, e.g. with \code{\link[=use_github_action]{use_github_action()}}.
}
\keyword{internal}
