% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{use_travis_deploy}
\alias{use_travis_deploy}
\title{Setup deployment for Travis CI}
\usage{
use_travis_deploy()
}
\description{
Creates a public-private key pair,
adds the public key to the GitHub repository via \code{\link[travis:github_add_key]{travis::github_add_key()}},
and stores the private key as an encrypted environment variable in Travis CI
via \code{\link[travis:travis_set_var]{travis::travis_set_var()}},
possibly in a different repository.
The \pkg{tic} companion package contains facilities for installing such a key
during a Travis CI build.
}
