% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ui.R
\name{use_ui}
\alias{use_ui}
\title{Use usethis'ui(s) in your package}
\usage{
use_ui()
}
\description{
Does setup necessary to use the usethis' user interfaces in your package.
This function requires the use roxygen.
\itemize{
\item Check that the active package uses roxygen2
\item Adds usethis package to "Imports" in \code{DESCRIPTION}
\item Imports in your namespace:
\itemize{
\item block styles: \code{\link{ui_line}}, \code{\link{ui_todo}}
\code{\link{ui_done}}, \code{\link{ui_todo}}, \code{\link{ui_oops}}
\code{\link{ui_info}}, \code{\link{ui_code_block}}
\item conditions: \code{\link{ui_stop}}, \code{\link{ui_warn}}
\item questions: \code{\link{ui_yeah}}, \code{\link{ui_nope}}
\item inline styles: \code{\link{ui_field}}, \code{\link{ui_value}}
\code{\link{ui_path}}, \code{\link{ui_code}}) user interfaces
}
}
}
\examples{
\dontrun{
use_ui()
}
}
