% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{use_version}
\alias{use_version}
\title{Increment package version}
\usage{
use_version(which = NULL)
}
\arguments{
\item{which}{A string specifying which level to increment, one of: "major",
"minor", "patch", "dev". If \code{NULL}, user can choose interactively.}
}
\description{
Increments the "Version" field in \code{DESCRIPTION}, adds a new heading to
\code{NEWS.md} (if it exists), and commits those changes (if package uses Git).
}
