% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{usethis-defunct}
\alias{usethis-defunct}
\alias{use_depsy_badge}
\alias{pr_pull_upstream}
\title{Defunct functions in usethis}
\usage{
use_depsy_badge()

pr_pull_upstream()
}
\description{
These functions are marked as defunct and have been removed from usethis.
}
\keyword{internal}
