% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-package.R
\docType{package}
\name{usethis-package}
\alias{usethis}
\alias{usethis-package}
\title{usethis: Automate Package and Project Setup}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automate package and project setup tasks that are otherwise performed manually. This includes setting up unit testing, test coverage, continuous integration, Git, 'GitHub', licenses, 'Rcpp', 'RStudio' projects, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usethis.r-lib.org}
  \item \url{https://github.com/r-lib/usethis}
  \item Report bugs at \url{https://github.com/r-lib/usethis/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
  \item Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})
  \item Andy Teucher \email{andy.teucher@posit.co} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
