% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_union}
\alias{write_union}
\title{Write into a file}
\usage{
write_union(path, new_lines, quiet = FALSE)
}
\arguments{
\item{path}{Path to target file. It is created if it does not exist, but the
parent directory must exist.}

\item{new_lines}{Character vector of lines to add, if not already present.}

\item{quiet}{Logical. Whether to message about what is happening.}
}
\value{
Logical indicating whether a write occurred, invisibly.
}
\description{
Writes lines to a file, taking the union of what's already there, if
anything, and some new lines. Note, there is no explicit promise about the
line order. This function is designed to modify simple config files like
\code{.Rbuildignore} and \code{.gitignore}.
}
\examples{
tmp <- tempfile()
write_union(tmp, letters[1:3])
readLines(tmp)
write_union(tmp, letters[1:5])
readLines(tmp)

## clean up
file.remove(tmp)
}
\keyword{internal}
