% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{write_vscode_json}
\alias{write_vscode_json}
\title{Write JSON to a VS Code settings file}
\usage{
write_vscode_json(x, path)
}
\description{
Small shim to use in place of \code{\link[jsonlite:read_json]{jsonlite::write_json()}} when writing to
\code{.vscode/settings.json} or \code{.vscode/extensions.json}.

Notably:
\itemize{
\item 4 space indent, as that is the standard indent level for these files
\item Auto unbox, because we want \code{TRUE} to show up as \code{true} not \verb{[true]}.
\item Trims newlines from the right hand side after the ending \verb{\}}. Unfortunately
setting \code{pretty = 4L} causes the special libyajl formatter to kick in, and
that always adds a trailing newline after every \verb{]} or \verb{\}}, even the last
one, which we don't want.
}
}
\keyword{internal}
