% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{use_ci}
\alias{use_ci}
\alias{use_tic}
\title{Continuous integration setup and badges}
\usage{
use_ci(path = ".", quiet = FALSE, services = c("Travis", "Appveyor"))

use_tic(repo_type)
}
\arguments{
\item{path}{\code{[string]}\cr
The path to the repo to prepare.}

\item{quiet}{\code{[flag]}\cr
Less verbose output? Default: \code{FALSE}.}

\item{services}{\code{[character]}\cr
CI services to add.}

\item{repo_type}{The kind of project for which a tic.R file should be created.
This is automatically guessed when using \link{use_ci}. Valid values are \code{blogdown},
\code{bookdown}, \code{package} and \code{site}.}
}
\description{
Sets up continuous integration (CI) services for an R package that is
developed on GitHub with CI-agnostic support by \pkg{tic}.
CI services can run \code{R CMD check} automatically on
various platforms, triggered by each push or pull request. This function
\itemize{
\item Adds service-specific configuration files and adds them to \code{.Rbuildignore}.
\item Activates a service or gives the user a detailed prompt.\cr
}
}
\details{
By default the CI-services "Travis" (Linux) and "Appveyor"
(Windows) are used. Basic \code{.travis.yml} and \code{appveyor.yml} files are added
to the top-level directory of a package.

The following steps will be run:
\enumerate{
\item If necessary, create a GitHub repository via \code{\link[=use_github]{use_github()}}
\item Enable Travis via \code{\link[=travis_enable]{travis_enable()}}
\item Create a default \code{.travis.yml} file
(overwrite after confirmation in interactive mode only)
\item Create a default \code{appveyor.yml} file
(depending on repo type, overwrite after confirmation
in interactive mode only)
\item Create a default \code{tic.R} file depending on the repo type
(package, website, bookdown, ...)
\item Enable deployment (if necessary, depending on repo type)
via \code{\link[=use_travis_deploy]{use_travis_deploy()}}
\item Create a GitHub PAT and install it on Travis CI via \code{\link[=travis_set_pat]{travis_set_pat()}}
}

This function is aimed at supporting the most common use cases.
Users who require more control are advised to manually call the individual
functions.
}
