% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{use_travis}
\alias{use_travis_badge}
\alias{use_appveyor}
\alias{use_appveyor_badge}
\alias{use_gitlab_ci}
\alias{use_circleci}
\alias{use_circleci_badge}
\alias{use_azure_pipelines}
\alias{use_github_actions}
\alias{use_github_actions_badge}
\alias{use_action}
\title{Continuous integration setup and badges}
\usage{
use_travis(browse = interactive(), ext = c("org", "com"))

use_travis_badge(ext = "org")

use_appveyor(browse = interactive())

use_appveyor_badge()

use_gitlab_ci()

use_circleci(browse = interactive(), image = "rocker/verse:latest")

use_circleci_badge()

use_azure_pipelines(browse = interactive())

use_github_actions()

use_github_actions_badge(name = "R")

use_action(name,
  url = glue("https://raw.githubusercontent.com/r-lib/actions/master/examples/{name}"),
  save_as = basename(url), ignore = FALSE, open = FALSE)
}
\arguments{
\item{browse}{Open a browser window to enable automatic builds for the
package.}

\item{ext}{which travis website to use. default to \code{"org"}for
https://travis-ci.org. Change to \code{"com"} for https://travis-ci.com.}

\item{image}{The Docker image to use for build. Must be available on
\href{https://hub.docker.com}{DockerHub}. The
\href{https://hub.docker.com/r/rocker/verse}{rocker/verse} image includes TeX
Live, pandoc, and the tidyverse packages. For a minimal image, try
\href{https://hub.docker.com/r/rocker/r-ver}{rocker/r-ver}. To specify a version
of R, change the tag from \code{latest} to the version you want, e.g.
\code{rocker/r-ver:3.5.3}.}

\item{name}{The name to give to the GitHub Actions workflow}

\item{url}{The full URL to the action yaml file.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{ignore}{Should the newly created file be added to \code{.Rbuildignore}?}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Sets up continuous integration (CI) services for an R package that is
developed on GitHub. CI services can run \code{R CMD check} automatically on
various platforms, triggered by each push or pull request. These functions
\itemize{
\item Add service-specific configuration files and add them to \code{.Rbuildignore}.
\item Activate a service or give the user a detailed prompt.
\item Provide the markdown to insert a badge into README.
}
}
\section{\code{use_travis()}}{

Adds a basic \code{.travis.yml} to the top-level directory of a package. This is a
configuration file for the \href{https://travis-ci.org/}{Travis CI} continuous
integration service.
}

\section{\code{use_travis_badge()}}{

Only adds the \href{https://travis-ci.org/}{Travis CI} badge. Use for a project
where Travis is already configured.
}

\section{\code{use_appveyor()}}{

Adds a basic \code{appveyor.yml} to the top-level directory of a package. This is
a configuration file for the \href{https://www.appveyor.com}{AppVeyor} continuous
integration service for Windows.
}

\section{\code{use_appveyor_badge()}}{

Only adds the \href{https://www.appveyor.com}{AppVeyor} badge. Use for a project
where AppVeyor is already configured.
}

\section{\code{use_gitlab_ci()}}{

Adds a basic \code{.gitlab-ci.yml} to the top-level directory of a package. This is
a configuration file for the \href{https://docs.gitlab.com/ee/ci/}{GitLab CI/CD} continuous
integration service for GitLab.
}

\section{\code{use_circleci()}}{

Adds a basic \code{.circleci/config.yml} to the top-level directory of a package. This is a
configuration file for the \href{https://circleci.com/}{CircleCI} continuous
integration service.
}

\section{\code{use_circleci_badge()}}{

Only adds the \href{https://www.circleci.com}{Circle CI} badge. Use for a project
where Circle CI is already configured.
}

\section{\code{use_azure_pipelines()}}{

Adds a basic \code{azure-pipelines.yml} to the top-level directory of a package. This is
a configuration file for the \href{https://azure.microsoft.com/en-us/services/devops/pipelines/}{Azure Pipelines} continuous
integration service.
}

\section{\code{use_github_actions()}}{

Adds a basic ‘r.yaml’ file to the \code{.github/workflows} directory of a
package. This is a configuration file for the GitHub Actions service'
}

\section{\code{use_github_actions_badge()}}{

Only adds the \href{https://github.com/features/actions}{GitHub Actions} badge. Use for a project
where GitHub actions is already configured.
}

\section{\code{use_action()}}{

Use a specific action, either one fo the example actions from
\href{https://github.com/r-lib/actions/tree/master/examples}{@r-lib/actions/examples} or a custom action
given by the \code{url} parameter.
}

