% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_from_github}
\alias{create_from_github}
\title{Create a project from a github repository}
\usage{
create_from_github(repo, path = ".", fork = NA, open = TRUE)
}
\arguments{
\item{repo}{Full name of repository: \code{owner/repo}}

\item{path}{Directory in which to clone repository: will create new
directory inside \code{path}.}

\item{fork}{Create a fork before cloning? Defaults to \code{TRUE} if you
can't push to \code{repo}, \code{FALSE} if you can.}

\item{open}{Open the new project once cloned?}
}
\description{
Create a project from a github repository
}
