% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{edit}
\alias{edit}
\alias{edit_profile_user}
\alias{edit_environ_user}
\alias{edit_makevars_user}
\alias{edit_git_config_user}
\alias{edit_git_ignore_user}
\title{Open useful configuration files}
\usage{
edit_profile_user()

edit_environ_user()

edit_makevars_user()

edit_git_config_user()

edit_git_ignore_user()
}
\description{
\itemize{
\item \code{edit_profile_user()} opens \code{~/.Rprofile}
\item \code{edit_environ_user()} opens \code{~/.Renviron}
\item \code{edit_makevars_users()} opens \code{~/.R/Makevars}
\item \code{edit_git_config_user()} opens \code{~/.gitconfig}
\item \code{edit_git_ignore_user()} opens \code{~/.gitignore}
}
}
