% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-deprecated.R
\name{git_branch_default}
\alias{git_branch_default}
\title{Deprecated Git functions}
\usage{
git_branch_default()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{git_branch_default()} has been replaced by \code{\link[=git_default_branch]{git_default_branch()}}.
}
\keyword{internal}
