% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_branch_default}
\alias{git_branch_default}
\title{Determine default Git branch}
\usage{
git_branch_default()
}
\value{
A branch name
}
\description{
Figure out the default branch of the current Git repo.
}
\examples{
\dontrun{
git_branch_default()
}
}
