% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{git_credentials}
\alias{git_credentials}
\alias{use_git_credentials}
\title{Produce or register credentials for git2r}
\usage{
git_credentials(protocol = deprecated(), auth_token = deprecated())

use_git_credentials(credentials = deprecated())
}
\arguments{
\item{protocol}{Deprecated.}

\item{auth_token}{Deprecated.}

\item{credentials}{Deprecated.}
}
\value{
These functions raise a warning and return an invisible \code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

In usethis v2.0.0, usethis switched from git2r to gert (+ credentials) for
all Git operations. This pair of packages (gert + credentials) is designed to
discover and use the same credentials as command line Git. As a result, a
great deal of credential-handling assistance has been removed from usethis,
primarily around SSH keys.

If you have credential problems, focus your troubleshooting on getting the
credentials package to find your credentials. The \href{https://docs.ropensci.org/credentials/articles/intro.html}{introductory vignette}
is a good place to start.

If you use the HTTPS protocol (which we recommend), a GitHub personal access
token will satisfy all auth needs, for both Git and the GitHub API, and is
therefore the easiest approach to get working. See \code{\link[=gh_token_help]{gh_token_help()}} for
more.
}
