% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_protocol}
\alias{git_protocol}
\alias{use_git_protocol}
\title{Produce or register Git protocol}
\usage{
git_protocol()

use_git_protocol(protocol)
}
\arguments{
\item{protocol}{Optional. Should be "https" or "ssh", if specified. Defaults
to the option \code{usethis.protocol} and, if unset, to an interactive choice
or, in non-interactive sessions, \code{"https"}. \code{NA} triggers the interactive
menu.}
}
\value{
"https" or "ssh"
}
\description{
Git operations that address a remote use a so-called "transport protocol".
usethis supports HTTPS and SSH. The protocol affects this:
\itemize{
\item The default URL format for repos with no existing remote protocol:
\itemize{
\item \code{protocol = "https"} implies \verb{https://github.com/<OWNER>/<REPO>.git}
\item \code{protocol = "ssh"} implies \verb{git@github.com:<OWNER>/<REPO>.git}
Two helper functions are available:
}
\item \code{git_protocol()} returns the user's preferred protocol, if known, and,
otherwise, asks the user (interactive session), or defaults to \code{"https"}.
(non-interactive session).
\item \code{use_git_protocol()} allows the user to set the Git protocol for the
current R session. This is stored in the \code{"usethis.protocol"} option.
}

Any interactive choice re: \code{protocol} comes with a reminder of how to set the
protocol at startup by setting an option in \code{.Rprofile}:\preformatted{options(usethis.protocol = "https")
# or
options(usethis.protocol = "ssh")
}
}
\examples{
\dontrun{
## consult the option and maybe get an interactive menu
git_protocol()

## explicitly set the protocol
use_git_protocol("ssh")
use_git_protocol("https")
}
}
