% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_sitrep}
\alias{git_sitrep}
\title{git/GitHub sitrep}
\usage{
git_sitrep()
}
\description{
Get a situation report on your current git/GitHub status. Useful for
diagnosing problems.
}
\examples{
git_sitrep()
}
