% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_sitrep}
\alias{git_sitrep}
\title{Git/GitHub sitrep}
\usage{
git_sitrep(tool = c("git", "github"), scope = c("user", "project"))
}
\arguments{
\item{tool}{Report for \strong{git}, or \strong{github}}

\item{scope}{Report globally for the current \strong{user}, or locally for the
current \strong{project}}
}
\description{
Get a situation report on your current Git/GitHub status. Useful for
diagnosing problems. The default is to report all values; provide values
for \code{tool} or \code{scope} to be more specific.
}
\examples{
\dontrun{
# report all
git_sitrep()

# report git for current user
git_sitrep("git", "user")
}
}
