% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_vaccinate}
\alias{git_vaccinate}
\title{Vaccinate your global gitignore file}
\usage{
git_vaccinate()
}
\description{
Adds \code{.DS_Store}, \code{.Rproj.user}, \code{.Rdata}, \code{.Rhistory}, and \code{.httr-oauth} to
your global (a.k.a. user-level) \code{.gitignore}. This is good practice as it
decreases the chance that you will accidentally leak credentials to GitHub.
}
