% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-actions.R
\name{github_actions}
\alias{github_actions}
\alias{use_github_actions}
\alias{use_github_actions_badge}
\title{GitHub Actions setup and badges}
\usage{
use_github_actions()

use_github_actions_badge(name = "R-CMD-check", repo_spec = NULL)
}
\arguments{
\item{name}{The name of an existing \href{https://github.com/features/actions}{GitHub Actions} workflow.}

\item{repo_spec}{GitHub repo specification in this form: \code{owner/repo}.
Default is to infer from GitHub remotes of active project.}
}
\description{
Sets up continuous integration (CI) for an R package that is developed on
GitHub using \href{https://github.com/features/actions}{GitHub Actions}. These
functions
\itemize{
\item Add the necessary configuration files and list them in \code{.Rbuildignore}.
\item Provide the markdown to insert a badge into your README
}
}
\section{\code{use_github_actions()}}{

Adds a basic \code{R-CMD-check.yaml} file to the \code{.github/workflows} directory of
a package. This is a configuration file for the \href{https://github.com/features/actions}{GitHub Actions} service.
}

\section{\code{use_github_actions_badge()}}{

Only adds the \href{https://github.com/features/actions}{GitHub Actions} badge.
Use for a project where GitHub Actions is already configured.
}

\seealso{
\code{\link[=use_github_action]{use_github_action()}} for setting up a specific action.
}
