% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-actions.R
\name{github_actions}
\alias{github_actions}
\alias{use_github_actions}
\alias{use_github_actions_tidy}
\alias{use_github_actions_badge}
\alias{use_github_action}
\alias{use_github_action_check_release}
\alias{use_github_action_check_full}
\alias{use_github_action_pr_commands}
\title{GitHub Actions setup and badges}
\usage{
use_github_actions()

use_github_actions_tidy()

use_github_actions_badge(name = "R-CMD-check")

use_github_action(
  name,
  url = glue("https://raw.githubusercontent.com/r-lib/actions/master/examples/{name}"),
  save_as = basename(url),
  ignore = TRUE,
  open = FALSE
)

use_github_action_check_release(
  save_as = "R-CMD-check.yaml",
  ignore = TRUE,
  open = FALSE
)

use_github_action_check_full(
  save_as = "R-CMD-check.yaml",
  ignore = TRUE,
  open = FALSE
)

use_github_action_pr_commands(
  save_as = "pr-commands.yaml",
  ignore = TRUE,
  open = FALSE
)
}
\arguments{
\item{name}{The name to give to the \href{https://github.com/features/actions}{GitHub Actions} workflow.}

\item{url}{The full URL to the GitHub Actions yaml file.}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to \code{template}}

\item{ignore}{Should the newly created file be added to \code{.Rbuildignore}?}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Sets up continuous integration (CI) for an R package that is developed on
GitHub using \href{https://github.com/features/actions}{GitHub Actions}. These functions
\itemize{
\item Add the necessary configuration files and place them in \code{.Rbuildignore}.
\item Provide the markdown to insert a badge into your README
}
}
\section{\code{use_github_actions()}}{

Adds a basic ‘R-CMD-check.yaml’ file to the \code{.github/workflows} directory of a
package. This is a configuration file for the \href{https://github.com/features/actions}{GitHub Actions} service.
}

\section{\code{use_github_actions_tidy()}}{

Sets up tidyverse actions that check the R versions officially supported by
the tidyverse, (current release, devel and four previous versions). It also
adds two commands to be used in pull requests, \verb{\\document} to run
\code{roxygen2::roxygenise()} and update the PR, and \verb{\\style} to run
\code{styler::style_pkg()} and update the PR.
}

\section{\code{use_github_actions_badge()}}{

Only adds the \href{https://github.com/features/actions}{GitHub Actions} badge. Use for a project
where GitHub Actions is already configured.
}

\section{\code{use_github_action()}}{

Use a specific action, either one of the example actions from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions/examples} or a custom action
given by the \code{url} parameter.
}

\section{\code{use_github_action_check_release()}}{

This action installs the latest release R version on macOS and runs \verb{R CMD check}
via the \href{https://github.com/r-lib/rcmdcheck}{rcmdcheck} package.
}

\section{\code{use_github_action_check_full()}}{

This action installs the last 5 minor R versions and runs R CMD check
via the \href{https://github.com/r-lib/rcmdcheck}{rcmdcheck} package on the
three major OSs (linux, macOS and Windows). This action is what the
tidyverse teams uses on their repositories, but is overkill for less
widely used packages, which are better off using the simpler
\code{use_github_action_check_release()}.
}

\section{\code{use_github_action_pr_commands()}}{

This workflow enables the use of 2 R specific commands in pull request
issue comments. \verb{\\document} will use
\href{https://roxygen2.r-lib.org/}{roxygen2} to rebuild the documentation for
the package and commit the result to the pull request. \verb{\\style} will use
\href{https://styler.r-lib.org/}{styler} to restyle your package.
}

