% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_token.R
\name{github_lock_branch}
\alias{github_lock_branch}
\alias{github_unlock_branch}
\title{Lock and unlock a branch on GitHub}
\usage{
github_lock_branch(branch = NULL)

github_unlock_branch(branch = NULL)
}
\arguments{
\item{branch}{The branch to lock/unlock. If not supplied, uses the
default branch which is usually "main" or "master".}
}
\description{
These functions lock and unlock a branch on GitHub so that it's not possible
for anyone to make any changes. This is used as part of the release process
to ensure that you don't accidentally merge any pull requests or push any
commits while you are waiting for CRAN to get back to you.

You must be an admin or an owner of the repo in order to lock/unlock
a branch.
}
