% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license.R
\name{licenses}
\alias{use_gpl3_license}
\alias{use_mit_license}
\title{Add licenses}
\usage{
use_mit_license(pkg = ".", copyright_holder = getOption("devtools.name",
  "<Author>"))

use_gpl3_license(pkg = ".")
}
\arguments{
\item{copyright_holder}{The copyright holder for this package. Defaults to
\code{getOption("devtools.name")}.}
}
\description{
Add licenses
}
\section{\code{use_mit_license}}{

Adds the necessary infrastructure to declare your package as
distributed under the MIT license.
}

\section{\code{use_gpl3_license}}{

Adds the necessary infrastructure to declare your package as
distributed under the GPL v3.
}

