% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr_init}
\alias{pr_init}
\alias{pr_fetch}
\alias{pr_push}
\alias{pr_pull}
\alias{pr_view}
\alias{pr_finish}
\title{Helpers for GitHub pull requests}
\usage{
pr_init(branch)

pr_fetch(number)

pr_push()

pr_pull()

pr_view()

pr_finish()
}
\arguments{
\item{branch}{branch name. Should usually consist of lower case letters,
numbers, and \code{-}.}

\item{number}{Number of PR to fetch.}
}
\description{
\itemize{
\item \code{pr_init("name")} creates a new local branch for a PR.
\item \code{pr_fetch(number)} downloads a remote PR so you can edit locally.
\item \code{pr_push()} pushes local changes to GitHub, after checking that there
aren't any remote changes you need to retrieve first. On first use,
it will prompt you to create a PR on GitHub.
\item \code{pr_pull()} retrives changes from the GitHub PR. Run this if others
have made suggestions or pushed into your PR.
\item \code{pr_view()} open the PR in the browser
}
}
\details{
These functions have been designed to support the git and GitHub best
practices described in \url{http://happygitwithr.com/}.
}
