% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_activate}
\alias{proj_activate}
\title{Activate a project}
\usage{
proj_activate(path)
}
\arguments{
\item{path}{Project directory}
}
\value{
Single logical value indicating if current session is modified.
}
\description{
Activates a project in usethis, R session, and (if relevant) RStudio senses.
If you are in RStudio, this will open a new RStudio session. If not, it will
change the working directory and \link[=proj_set]{active project}.
}
