% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{proj_sitrep}
\alias{proj_sitrep}
\title{Report working directory and usethis/RStudio project}
\usage{
proj_sitrep()
}
\value{
A named list, with S3 class \code{sitrep} (for printing purposes),
reporting current working directory, active usethis project, and active
RStudio Project
}
\description{
Report working directory and usethis/RStudio project
}
\examples{
proj_sitrep()
}
