% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{proj_sitrep}
\alias{proj_sitrep}
\title{Report working directory and usethis/RStudio project}
\usage{
proj_sitrep()
}
\value{
A named list, with S3 class \code{sitrep} (for printing purposes),
reporting current working directory, active usethis project, and active
RStudio Project
}
\description{
\code{proj_sitrep()} reports
\itemize{
\item current working directory
\item the active usethis project
\item the active RStudio Project
}

Call this function if things seem weird and you're not sure
what's wrong or how to fix it. Usually, all three of these should coincide
(or be unset) and \code{proj_sitrep()} provides suggested commands for getting
back to this happy state.
}
\examples{
proj_sitrep()
}
\seealso{
Other project functions: 
\code{\link{proj_utils}}
}
\concept{project functions}
