% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{rprofile-helper}
\alias{rprofile-helper}
\alias{use_conflicted}
\alias{use_reprex}
\alias{use_usethis}
\alias{use_devtools}
\alias{use_partial_warnings}
\title{Helpers to make useful changes to \code{.Rprofile}}
\usage{
use_conflicted()

use_reprex()

use_usethis()

use_devtools()

use_partial_warnings()
}
\description{
All functions open your \code{.Rprofile} and give you the code you need to
paste in.
\itemize{
\item \code{use_devtools()}: makes devtools available in interactive sessions.
\item \code{use_usethis()}: makes usethis available in interactive sessions.
\item \code{use_reprex()}: makes reprex available in interactive sessions.
\item \code{use_conflicted()}:  makes conflicted available in interactive sessions.
\item \code{use_partial_warnings()}: warns on partial matches.
}
}
