% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{tidyverse}
\alias{tidyverse}
\alias{use_tidy_ci}
\title{Helpers for the tidyverse}
\usage{
use_tidy_ci(browse = interactive(), base_path = ".")
}
\arguments{
\item{browse}{open a browser window to enable Travis builds for the package
automatically.}

\item{base_path}{Path to package root.}
}
\description{
These helpers follow tidyverse conventions which are generally a little
stricter than the defaults, reflecting the need for greater rigor in
commonly used packages
}
\details{
\itemize{
\item \code{use_tidy_ci()}: sets up travis and codecov
}
}
