% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{tidyverse}
\alias{tidyverse}
\alias{use_tidy_ci}
\alias{use_tidy_description}
\alias{use_tidy_eval}
\title{Helpers for the tidyverse}
\usage{
use_tidy_ci(browse = interactive())

use_tidy_description()

use_tidy_eval()
}
\arguments{
\item{browse}{open a browser window to enable Travis builds for the package
automatically.}
}
\description{
These helpers follow tidyverse conventions which are generally a little
stricter than the defaults, reflecting the need for greater rigor in
commonly used packages
}
\details{
\itemize{
\item \code{use_tidy_ci()}: sets up travis and codecov, ensuring that the package
works on all version of R starting at 3.1.
\item \code{use_tidy_description()}: puts fields in standard order and alphabetises
dependencies.
\item \code{use_tidy_eval()}: imports a standard set of helpers to facilitate
programming with the tidy eval toolkit.
}
}
