% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{use_coverage}
\alias{use_coverage}
\title{Add test coverage via codecov or coveralls}
\usage{
use_coverage(type = c("codecov", "coveralls"))
}
\arguments{
\item{type}{Which web service to use for test reporting. Currently supports
\href{https://codecov.io}{Codecov} and \href{https://coveralls.io}{Coveralls}.}
}
\description{
Enables test coverage reporting via codecov or coveralls
}
\section{\code{use_coverage()}}{

Adds test coverage reports to a package that is already using Travis CI.
}

