% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{use_coverage}
\alias{use_coverage}
\alias{use_covr_ignore}
\title{Test coverage}
\usage{
use_coverage(type = c("codecov", "coveralls"), repo_spec = NULL)

use_covr_ignore(files)
}
\arguments{
\item{type}{Which web service to use.}

\item{repo_spec}{Optional GitHub repo specification in this form: \code{owner/repo}. This can usually be inferred from the GitHub remotes of active project.}

\item{files}{Character vector of file globs.}
}
\description{
Adds test coverage reporting to a package, using either Codecov
(\verb{https://codecov.io}) or Coveralls (\verb{https://coveralls.io}).
}
