% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{use_data}
\alias{use_data}
\alias{use_data_raw}
\title{Create package data}
\usage{
use_data(..., internal = FALSE, overwrite = FALSE, compress = "bzip2",
  base_path = ".")

use_data_raw(base_path = ".")
}
\arguments{
\item{...}{Unquoted names of existing objects to save.}

\item{internal}{If \code{FALSE}, saves each object in individual
\code{.rda} files in the \code{data/} directory. These are available
whenever the package is loaded. If \code{TRUE}, stores all objects in
a single \code{R/sysdata.rda} file. These objects are only available
within the package.}

\item{overwrite}{By default, \code{use_data} will not overwrite existing
files. If you really want to do so, set this to \code{TRUE}.}

\item{compress}{Choose the type of compression used by \code{\link{save}}.
Should be one of "gzip", "bzip2" or "xz".}

\item{base_path}{Path to package root.}
}
\description{
\code{use_data} makes it easy to save package data in the correct format.
I recommend you save scripts that generate package data in \code{data-raw}:
use \code{use_data_raw} to set it up.
}
\examples{
\dontrun{
x <- 1:10
y <- 1:100

use_data(x, y) # For external use
use_data(x, y, internal = TRUE) # For internal use
}
}

