% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{use_data_raw}
\alias{use_data_raw}
\title{Use \code{data-raw} to compute package datasets.}
\usage{
use_data_raw(pkg = ".")
}
\arguments{
\item{pkg}{Package where to create \code{data-raw}. Defaults to package in
working directory.}
}
\description{
Use \code{data-raw} to compute package datasets.
}
\seealso{
Other infrastructure: \code{\link{use_build_ignore}},
  \code{\link{use_data}}, \code{\link{use_news_md}},
  \code{\link{use_package}}, \code{\link{use_readme_rmd}},
  \code{\link{use_travis}}
}

