% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table.R
\name{use_data_table}
\alias{use_data_table}
\title{Prepare for importing data.table}
\usage{
use_data_table()
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#questioning'><img src='figures/lifecycle-questioning.svg' alt='Questioning lifecycle'></a>}}{\strong{Questioning}}

\code{use_data_table} facilitates importing the data.table package by handling
up-front some common set-up tasks for using it in your package.

This function does two main things:
\enumerate{
\item Import the entire data.table namespace (with \verb{@import}).
\item Block the usage of data.table as a dependency (\code{DESCRIPTION} field
\code{Depends}); \code{data.table} should be used as an \emph{imported} or \emph{suggested}
package only. See this
\href{https://github.com/Rdatatable/data.table/issues/3076}{discussion}.
}
}
