% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{use_env_var}
\alias{use_env_var}
\title{Set an environment variable in \code{.Renviron}}
\usage{
use_env_var(name, value = NULL, scope = c("user", "project"))
}
\arguments{
\item{name}{Name of the environment variable. Must contain only letters,
digits, and underscores, and must start with a letter or underscore.}

\item{value}{Value to set. By default, you are prompted to enter the value
securely using \code{\link[askpass:askpass]{askpass::askpass()}}. Otherwise, we \strong{do not recommend}
that you provide the value directly in code, as it may be visible in your
command history. This argument exists primarily to allow you to set an
environment variable to a value retrieved programmatically, e.g. from
\pkg{keyring}.}

\item{scope}{Edit globally for the current \strong{user} (\code{"user"}, the default),
or locally for the current \strong{project} (\code{"project"}). Setting a variable in
the user scope ensures that it is available in all future R sessions.}
}
\value{
Path to the \code{.Renviron} file, invisibly.
}
\description{
Adds or updates an environment variable in the user or project \code{.Renviron}
and immediately makes it available in the current session via \code{\link[=Sys.setenv]{Sys.setenv()}}.
}
\examples{
\dontrun{
use_env_var("OPENAI_API_KEY")
}
}
