% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{use_git}
\alias{use_git}
\title{Initialise a git repository.}
\usage{
use_git(message = "Initial commit")
}
\arguments{
\item{message}{Message to use for first commit.}
}
\description{
\code{use_git} initialises a git repository, adds important files
to \code{.gitignore}, and commits all files.
}
\examples{
\dontrun{use_git()}
}
\seealso{
Other git helpers: \code{\link{use_git_hook}},
  \code{\link{use_git_ignore}}
}
