% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-action.R
\name{use_github_action}
\alias{use_github_action}
\title{Use a specific GitHub action}
\usage{
use_github_action(
  name,
  url = NULL,
  save_as = NULL,
  ignore = TRUE,
  open = FALSE
)
}
\arguments{
\item{name}{Name of the GitHub action, with or without \code{.yaml} extension}

\item{url}{The full URL to the GitHub Actions yaml file.
By default, the corresponding action in https://github.com/r-lib/actions
will be used.}

\item{save_as}{Name of the actions file. Defaults to \code{basename(url)}.}

\item{ignore}{Should the newly created file be added to \code{.Rbuildignore}?}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Use a specific action, either one of the example actions from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions/examples} or a custom action
given by the \code{url} parameter.
}
\seealso{
\link{github_actions} for generic workflows.
}
