% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-actions.R
\name{use_github_actions}
\alias{use_github_actions}
\title{Use Github actions}
\usage{
use_github_actions()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{use_github_actions()} is deprecated because it was just an alias
for \code{\link[=use_github_action_check_release]{use_github_action_check_release()}}
}
\keyword{internal}
