% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-labels.R
\name{use_github_labels}
\alias{use_github_labels}
\alias{tidy_labels}
\title{Manage GitHub issue labels.}
\usage{
use_github_labels(labels = tidy_labels(), delete_default = FALSE,
  auth_token = NULL, host = NULL)

tidy_labels()
}
\arguments{
\item{labels}{Named character vector of labels. The names are the label text,
such as "bug", and the values are the label colours in hexadecimal, such as
"e02a2a". First, labels that don't yet exist are created, then label
colours are updated.}

\item{delete_default}{If \code{TRUE}, will remove GitHub default labels that do
not appear in the \code{labels} vector (presumably defaults that aren't relevant
to your workflow).}

\item{auth_token}{Provide a personal access token (PAT) from
\url{https://github.com/settings/tokens}. If \code{NULL}, will use the logic
described in \code{\link[gh:gh_whoami]{gh::gh_whoami()}} to look for a token stored in an environment
variable.}

\item{host}{GitHub API host to use. Override with the endpoint-root for your
GitHub enterprise instance, for example,
"https://github.hostname.com/api/v3". You can set this globally using the
\code{GITHUB_API_URL} env var.}
}
\description{
\code{use_github_labels()} creates new labels and/or changes label
colours. It does not generally remove labels. But if you set
\code{delete_default = TRUE}, it will delete labels that are (1) flagged by the
API as being \href{https://help.github.com/articles/about-labels/#using-default-labels}{GitHub default labels}
and (2) not present in the labels you provide via \code{labels}.

\code{tidy_labels()} returns the labels and colours commonly used by tidyverse
packages.
}
\examples{
\dontrun{
## typical use in, e.g., a new tidyverse project
use_github_labels(delete_default = TRUE)
}
}
