% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-labels.R
\name{use_github_labels}
\alias{use_github_labels}
\title{Create standard github labels for labelling issues.}
\usage{
use_github_labels(delete_default = FALSE, auth_token = NULL, host = NULL)
}
\arguments{
\item{delete_default}{If \code{TRUE}, will remove default labels.}

\item{auth_token}{Provide a personal access token (PAT) from
\url{https://github.com/settings/tokens}. If \code{NULL}, will use the
\code{GITHUB_PAT} environment variable.}

\item{host}{GitHub API host to use. Override with the endpoint-root for your
GitHub enterprise instance, for example,
"https://github.hostname.com/api/v3". You can set this globally using
the \code{GITHUB_API_URL} env var.}
}
\description{
This function creates new labels and changes colours (if needed).
It does not generally remove labels, unless you explicit ask to
remove GitHub's defaults.
}
