% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{use_github_release}
\alias{use_github_release}
\title{Publish a GitHub release}
\usage{
use_github_release(
  publish = TRUE,
  prerelease = FALSE,
  generate_release_notes = FALSE,
  host = deprecated(),
  auth_token = deprecated()
)
}
\arguments{
\item{publish}{If \code{TRUE}, publishes a release. If \code{FALSE}, creates a draft
release.}

\item{prerelease}{If \code{TRUE} identify the release as a prerelease.}

\item{generate_release_notes}{If \code{TRUE}, the body will be append with
automatically generated notes.}

\item{host, auth_token}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}: No longer
consulted now that usethis allows the gh package to lookup a token based on
a URL determined from the current project's GitHub remotes.}
}
\description{
Pushes the current branch (if safe) then publishes a GitHub
release for the latest CRAN submission.

If you use \code{\link[devtools:submit_cran]{devtools::submit_cran()}} to submit to CRAN, information about the
submitted state is captured in a \code{CRAN-SUBMISSION} file.
\code{use_github_release()} uses this info to populate the GitHub release notes
and, after success, deletes the file. In the absence of such a file, we
assume that current state (SHA of \code{HEAD}, package version, NEWS) is the
submitted state.
}
