% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{use_github_release}
\alias{use_github_release}
\title{Publish a GitHub release}
\usage{
use_github_release(publish = TRUE)
}
\arguments{
\item{publish}{If \code{TRUE}, publishes a release. If \code{FALSE}, creates a draft
release.}
}
\description{
Pushes the current branch (if safe) then publishes a GitHub release for the
latest CRAN submission.

If you use \code{\link[devtools:submit_cran]{devtools::submit_cran()}} to submit to CRAN, information about the
submitted state is captured in a \code{CRAN-SUBMISSION} file.
\code{use_github_release()} uses this info to populate the GitHub release notes
and, after success, deletes the file. In the absence of such a file, we
assume that current state (SHA of \code{HEAD}, package version, NEWS) is the
submitted state.
}
