% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest-dependencies.R
\name{use_latest_dependencies}
\alias{use_latest_dependencies}
\title{Use "latest" versions of all dependencies}
\usage{
use_latest_dependencies(overwrite = TRUE, source = c("CRAN", "local"))
}
\arguments{
\item{overwrite}{By default (\code{TRUE}), all dependencies will be modified.
Set to \code{FALSE} to only modify dependencies without version
specifications.}

\item{source}{Use "CRAN" or "local" package versions.}
}
\description{
Pins minimum versions of all \code{Imports} and \code{Depends} dependencies to latest
ones (as determined by \code{source}). Useful for the tidyverse package, but
should otherwise be used with extreme care.
}
\keyword{internal}
