% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest-dependencies.R
\name{use_latest_dependencies}
\alias{use_latest_dependencies}
\title{Use "latest" versions of all dependencies}
\usage{
use_latest_dependencies(overwrite = FALSE, source = c("local", "CRAN"))
}
\arguments{
\item{overwrite}{By default (\code{FALSE}), only dependencies without version
specifications will be modified. Set to \code{TRUE} to modify all dependencies.}

\item{source}{Use "local" or "CRAN" package versions.}
}
\description{
Pins minimum version of dependencies to latest (as determined by "source)"
Useful for tidyverse package, but should otherwise be used with extreme
care.
}
\keyword{internal}
