% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{use_lifecycle}
\alias{use_lifecycle}
\title{Use lifecycle badges}
\usage{
use_lifecycle()
}
\description{
This helper add lifecycle as a dependency, copies the lifecycle badges into
the \code{man/figures} folder, and reminds you how to use the badge syntax.

Learn more at \url{http://lifecycle.r-lib.org/}
}
\seealso{
\code{\link[=use_lifecycle_badge]{use_lifecycle_badge()}} to signal the \href{https://www.tidyverse.org/lifecycle/}{global lifecycle stage} of your package as a whole.
}
