% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{use_lifecycle}
\alias{use_lifecycle}
\title{Use lifecycle badges}
\usage{
use_lifecycle()
}
\description{
Call this to import the lifecycle badges and Rd macro into your
package.
\itemize{
\item The SVG badges are imported in \code{man/figures}.
\item The \code{RdMacros} field of the DESCRIPTION file is updated so you
can use the \verb{\\\\lifecycle\{\}} macro in your documentation.
}

See the \href{https://lifecycle.r-lib.org/articles/lifecycle.html}{getting started vignette} of the
lifecycle package.
}
\seealso{
\code{\link[=use_lifecycle_badge]{use_lifecycle_badge()}} to signal the \href{https://www.tidyverse.org/lifecycle/}{global lifecycle stage} of your package.
}
