% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{use_lifecycle}
\alias{use_lifecycle}
\title{Use lifecycle badges}
\usage{
use_lifecycle()
}
\description{
This helper:
\itemize{
\item Adds lifecycle as a dependency.
\item Imports \code{\link[lifecycle:deprecated]{lifecycle::deprecated()}} for use in function arguments.
\item Copies the lifecycle badges into \code{man/figures}.
\item Reminds you how to use the badge syntax.
}

Learn more at \url{https://lifecycle.r-lib.org/articles/communicate.html}
}
\seealso{
\code{\link[=use_lifecycle_badge]{use_lifecycle_badge()}} to signal the
\href{https://lifecycle.r-lib.org/articles/stages.html}{lifecycle stage} of
your package as whole
}
