% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linting.R
\name{use_lintr}
\alias{use_lintr}
\title{Use code linting}
\usage{
use_lintr(mode = c("local", "travis", "test"))
}
\arguments{
\item{mode}{Options are \code{"local"}, \code{"travis"}, or \code{"test"}.}
}
\description{
\code{use_lintr()} Adds support for code linting using the \code{lintr} package. THere are three linting modes available:
}
\details{
\itemize{
\item "local": local interactive linting only
\item "travis": non-failing, post-build linting on Travis CI
\item "test": testthat unit test
}
}
