% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{use_logo}
\alias{use_logo}
\title{README logo_}
\usage{
use_logo(img, geometry = "120x140")
}
\arguments{
\item{img}{The path to an existing image file}

\item{geometry}{a \link{geometry} string specifying area (for cropping) or size (for resizing).}
}
\description{
These helpers produce the markdown text you need in your README to include
package logos.
}
\examples{
\dontrun{
use_logo("usethis.png")
}
}
