% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{use_namespace}
\alias{use_namespace}
\title{Use a basic \code{NAMESPACE}}
\usage{
use_namespace()
}
\description{
This \code{NAMESPACE} exports everything, except functions that start
with a \code{.}.
}
\seealso{
The \href{https://r-pkgs.org/namespace.html}{namespace chapter} of
\href{https://r-pkgs.org}{R Packages}.
}
