% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{use_news_md}
\alias{use_news_md}
\title{Create a simple \code{NEWS.md}}
\usage{
use_news_md(base_path = ".", open = TRUE)
}
\arguments{
\item{base_path}{Path to package root.}

\item{open}{Should the new created file be opened in RStudio?}
}
\description{
This creates a basic \code{NEWS.md} in the root directory.
}
