% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{use_news_md}
\alias{use_news_md}
\title{Create a simple \code{NEWS.md}}
\usage{
use_news_md(open = rlang::is_interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
This creates a basic \code{NEWS.md} in the root directory.
}
\seealso{
The \href{https://r-pkgs.org/release.html#important-files}{important files section} of \href{https://r-pkgs.org}{R Packages}.
}
