% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{use_package_doc}
\alias{use_package_doc}
\title{Create package documentation template}
\usage{
use_package_doc()
}
\description{
Adds a roxygen template for package documentation. When you \code{document()}
roxygen will flesh out the Rd file using data from the \code{DESCRIPTION}.
That ensures you don't need to repeat the same information in multiple
places. This block is a good place to put global namespace tags like
\code{@importFrom}.
}
