% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{use_package_doc}
\alias{use_package_doc}
\title{Package-level documentation}
\usage{
use_package_doc(open = rlang::is_interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Adds a dummy \code{.R} file that will prompt roxygen to generate basic
package-level documentation. If your package is named "foo", this will make
help available to the user via \code{?foo} or \code{package?foo}. Once you call
\code{devtools::document()}, roxygen will flesh out the \code{.Rd} file using data from
the \code{DESCRIPTION}. That ensures you don't need to repeat the same information
in multiple places. This \code{.R} file is also a good place for roxygen
directives that apply to the whole package (vs. a specific function), such as
global namespace tags like \verb{@importFrom}.
}
\seealso{
The \href{https://r-pkgs.org/man.html}{documentation chapter} of \href{https://r-pkgs.org}{R Packages}
}
