% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{use_pipe}
\alias{use_pipe}
\title{Use magrittr's pipe in your package}
\usage{
use_pipe(export = TRUE)
}
\arguments{
\item{export}{If \code{TRUE}, the file \code{R/utils-pipe.R} is added, which provides
the roxygen template to import and re-export \verb{\%>\%}. If \code{FALSE}, the necessary
roxygen directive is added, if possible, or otherwise instructions are given.

#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The base R pipe has been available since R 4.1.0, which handles most cases
the \code{magrittr} pipe handles -- \verb{\%>\%} can usually just be replaced with \verb{|>}.
To read about the differences, read this \code{tidyverse}
\href{https://www.tidyverse.org/blog/2023/04/base-vs-magrittr-pipe/}{blog post}
describing some special cases where using the base R pipe requires handling
different from the \code{magrittr} pipe, including using
\itemize{
\item \code{x |> f(1, y = _)} instead of \code{x \%>\% f(1, y = .)},
\item \verb{x |> (\\(x) x[[1]]))()} instead of \code{x \%>\% .[[1]]},
\item \verb{x |> (\\(x) f(a = x, b = x))()} instead of \code{x \%>\% f(a = ., b = .)}, and
\item \code{x |> f()} instead of \code{x \%>\% f}.
}}
}
\description{
Does setup necessary to use magrittr's pipe operator, \verb{\%>\%} in your package.
This function requires the use of \pkg{roxygen2}.
\itemize{
\item Adds magrittr to "Imports" in \code{DESCRIPTION}.
\item Imports the pipe operator specifically, which is necessary for internal
use.
\item Exports the pipe operator, if \code{export = TRUE}, which is necessary to make
\verb{\%>\%} available to the users of your package.
}
}
\examples{
\dontrun{
use_pipe()
}
}
\keyword{internal}
