% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{use_pipe}
\alias{use_pipe}
\title{Use magrittr's pipe in your package}
\usage{
use_pipe()
}
\description{
This imports magrittr, and creates a \code{R/utils-pipe.R} with the necessary
roxygen template to import and re-export the pipe.
}
