% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{use_pipe}
\alias{use_pipe}
\title{Use magrittr's pipe in your package}
\usage{
use_pipe()
}
\description{
Does setup necessary to use magrittr's pipe internally in your package and to
re-export it for users of your package:
\itemize{
\item Adds magrittr to "Imports" in \code{DESCRIPTION}
\item Creates \code{R/utils-pipe.R} with the necessary roxygen template
}
}
