% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{use_pipe}
\alias{use_pipe}
\title{Use magrittr's pipe in your package}
\usage{
use_pipe(export = TRUE)
}
\arguments{
\item{export}{If \code{TRUE}, the file \code{R/utils-pipe.R} is added, which provides
the roxygen template to import and re-export \verb{\%>\%}. If \code{FALSE}, the necessary
roxygen directive is added, if possible, or otherwise instructions are given.}
}
\description{
Does setup necessary to use magrittr's pipe operator, \verb{\%>\%} in your package.
This function requires the use roxygen.
\itemize{
\item Adds magrittr to "Imports" in \code{DESCRIPTION}.
\item Imports the pipe operator specifically, which is necessary for internal
use.
\item Exports the pipe operator, if \code{export = TRUE}, which is necessary to make
\verb{\%>\%} available to the users of your package.
}
}
\examples{
\dontrun{
use_pipe()
}
}
