% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{use_pkgdown}
\alias{use_pkgdown}
\alias{use_pkgdown_travis}
\title{Use pkgdown}
\usage{
use_pkgdown(config_file = "_pkgdown.yml", destdir = "docs")

use_pkgdown_travis()
}
\arguments{
\item{config_file}{Path to the pkgdown yaml config file}

\item{destdir}{Target directory for pkgdown docs}
}
\description{
\href{https://pkgdown.r-lib.org}{pkgdown} makes it easy to turn your package into
a beautiful website. There are two helper functions:
\itemize{
\item \code{use_pkgdown()}: creates a pkgdown config file and adds the file and
destination directory to \code{.Rbuildignore}.
\item \code{use_pkgdown_travis()}: helps you set up pkgdown for automatic deployment
on Travis-CI by:
\itemize{
\item Adding the \code{docs/} folder to \code{.gitignore} and \code{.Rbuildignore}.
\item Optionally building favicons if your package contains a package logo,
and adding the resulting \code{pkgdown/} folder to \code{.Rbuildignore}.
\item Creating an empty \code{gh-pages} branch for the pkgdown site to be
deployed to.
\item Prompting you about what to do next regarding Travis-CI deployment
keys and updating your \code{.travis.yml}.
}
}
}
\seealso{
\url{https://pkgdown.r-lib.org/articles/pkgdown.html#configuration}
}
