% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{use_pkgdown}
\alias{use_pkgdown}
\title{Use pkgdown}
\usage{
use_pkgdown()
}
\description{
\href{https://github.com/hadley/pkgdown}{pkgdown} makes it easy to turn your
package into a beautiful website. This helper creates \code{_pkgdown.yml}
and \code{docs/} for you, and adds them to \code{.Rbuildignore}
}
