% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{use_r}
\alias{use_r}
\title{Create or edit a .R file}
\usage{
use_r(name = NULL)
}
\arguments{
\item{name}{File name, without extension; will create if it doesn't already
exist. If not specified, and you're currently in a test file, will guess
name based on test name.}
}
\description{
Create or edit a .R file
}
\seealso{
\code{\link[=use_test]{use_test()}}, and also the \href{http://r-pkgs.had.co.nz/r.html}{R code chapter} of \href{http://r-pkgs.had.co.nz}{R Packages}.
}
