% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp.R
\name{use_rcpp}
\alias{use_rcpp}
\alias{use_rcpp_armadillo}
\alias{use_rcpp_eigen}
\alias{use_c}
\title{Use C, C++, RcppArmadillo, or RcppEigen}
\usage{
use_rcpp(name = NULL)

use_rcpp_armadillo(name = NULL)

use_rcpp_eigen(name = NULL)

use_c(name = NULL)
}
\arguments{
\item{name}{If supplied, creates and opens \verb{src/name.\{c,cpp\}}.}
}
\description{
Adds infrastructure commonly needed when using compiled code:
\itemize{
\item Creates \verb{src/}
\item Adds required packages to \code{DESCRIPTION}
\item May create an initial placeholder \code{.c} or \code{.cpp} file
\item Creates \code{Makevars} and \code{Makevars.win} files (\code{use_rcpp_armadillo()} only)
}
}
\details{
When using compiled code, please note that there must be at least one file
inside the \verb{src/} directory prior to building the package. As a result,
if an empty \verb{src/} directory is detected, either a \code{.c} or \code{.cpp} file will
be added.
}
