% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp.R
\name{use_rcpp}
\alias{use_rcpp}
\alias{use_rcpp_armadillo}
\alias{use_rcpp_eigen}
\alias{use_c}
\title{Use C, C++, RcppArmadillo, or RcppEigen}
\usage{
use_rcpp(name = NULL)

use_rcpp_armadillo(name = NULL)

use_rcpp_eigen(name = NULL)

use_c(name = NULL)
}
\arguments{
\item{name}{If supplied, creates and opens \code{src/name.{c,cpp}}.}
}
\description{
Creates \code{src/}, adds required packages to \code{DESCRIPTION},
optionally creates \code{.c} or \code{.cpp} files, and
where needed, \code{Makevars} and \code{Makevars.win} files.
}
