% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\alias{use_readme_md}
\alias{use_data_list_in_readme_rmd}
\alias{use_data_list_in_readme_md}
\title{Create README files.}
\usage{
use_readme_rmd(base_path = ".")

use_readme_md(base_path = ".")

use_data_list_in_readme_rmd(base_path = ".", stop_if_no_data = TRUE)

use_data_list_in_readme_md(base_path = ".", stop_if_no_data = TRUE)
}
\arguments{
\item{base_path}{Path to package root.}

\item{stop_if_no_data}{Logical. If \code{TRUE},
\code{use_data_list_in_readme_rmd} and \code{use_data_list_in_readme_md}
will throw an error if the package has no data directory, or the package
has not been installed. This is useful if you want to add datasets to the
package after creating the README.}
}
\description{
Creates skeleton README files with sections for
\itemize{
\item a high-level description of the package and its goals
\item R code to install from GitHub, if GitHub usage detected
\item a basic example
}
Use \code{Rmd} if you want a rich intermingling of code and data. Use
\code{md} for a basic README. \code{README.Rmd} will be automatically
added to \code{.Rbuildignore}. The resulting README is populated with default
YAML frontmatter and R fenced code blocks (\code{md}) or chunks (\code{Rmd}).
}
\details{
\code{use_data_list_in_readme_rmd} and \code{use_data_list_in_readme_md}
append a table to the README listing the datasets contained in the package.
}
\examples{
\dontrun{
use_readme_rmd()
use_readme_md()
}
}
