% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\alias{use_readme_md}
\alias{use_data_list_in_readme_rmd}
\alias{use_data_list_in_readme_md}
\title{Create README files.}
\usage{
use_readme_rmd(base_path = ".")

use_readme_md(base_path = ".")

use_data_list_in_readme_rmd(base_path = ".")

use_data_list_in_readme_md(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\description{
Creates skeleton README files with sections for
\itemize{
\item a high-level description of the package and its goals
\item R code to install from GitHub, if GitHub usage detected
\item a basic example
}
Use \code{Rmd} if you want a rich intermingling of code and data. Use
\code{md} for a basic README. \code{README.Rmd} will be automatically
added to \code{.Rbuildignore}. The resulting README is populated with default
YAML frontmatter and R fenced code blocks (\code{md}) or chunks (\code{Rmd}).
}
\details{
\code{use_data_list_in_readme_rmd} and \code{use_data_list_in_readme_md}
append a table to the README listing the datasets contained in the package.
}
\examples{
\dontrun{
# Examples not run, due to dependency on devtools

# Workflow 1: add data first, then create README
# 1. Create a package
pkgroot1 <- tempfile("testpkg")
devtools::create(pkgroot1)
# 2. Add the datasets
use_data(cars, base_path = pkgroot1)
# 3. Install the package
devtools::install(pkgroot1)
# 4. Check the you added all the datasets
list_datasets(basename(pkgroot1))
# 5. Create the README
use_readme_rmd(pkgroot1)
# or
use_readme_md(pkgroot1)

# Workflow 2: create README first, then add data
# 1. Create a package
pkgroot2 <- tempfile("testpkg")
devtools::create(pkgroot2)
# 2. Create the README
use_readme_rmd(pkgroot2)
# or
use_readme_md(pkgroot2)
# 3. Add the datasets
use_data(cars, base_path = pkgroot2)
# 4. Install the package
devtools::install(pkgroot2)
# 5. Check the you added all the datasets
list_datasets(basename(pkgroot2))
# 6. Update the README
use_data_list_in_readme_rmd(pkgroot2)
# or
use_data_list_in_readme_md(pkgroot2)
}
}
