% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep.R
\name{use_revdep}
\alias{use_revdep}
\title{Set up devtools revdep template}
\usage{
use_revdep()
}
\description{
Add \code{revdep} directory, git-ignoring the files that you shouldn't
check in, and creating a \code{email.yml} for use \code{revdep_email}.
Run the checks with \code{revdepcheck::revdep_check()}.
}
