% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep.R
\name{use_revdep}
\alias{use_revdep}
\title{Reverse dependency checks}
\usage{
use_revdep()
}
\description{
Performs set up for checking the reverse dependencies of an R package, as
implemented by the revdepcheck package:
\itemize{
\item Creates \verb{revdep/} directory and adds it to \code{.Rbuildignore}
\item Populates \code{revdep/.gitignore} to prevent tracking of various revdep
artefacts
\item Prompts user to run the checks with \code{revdepcheck::revdep_check()}
}
}
