% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{use_rstudio}
\alias{use_rstudio}
\title{Use RStudio}
\usage{
use_rstudio(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\description{
Creates an \code{.Rproj} file and adds RStudio files to \code{.gitignore}
and \code{.Rbuildignore}.
}

