% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{use_snapshot}
\alias{use_snapshot}
\title{Edit the snapshot file associated with an R or test file}
\usage{
use_snapshot(name = NULL, variant = NULL, open = rlang::is_interactive())
}
\arguments{
\item{name}{Either a string giving a file name (without directory) or
\code{NULL} to take the name from the currently open file in RStudio.}

\item{variant}{An optional subdirectory within the \verb{_snaps/} directory.}

\item{open}{Whether to open the file for interactive editing.}
}
\description{
This function opens the snapshot file associated with an R or test file, if
one exists.
}
