% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{use_spell_check}
\alias{use_spell_check}
\title{Use Spell Check}
\usage{
use_spell_check(vignettes = TRUE, lang = "en-US", error = FALSE)
}
\arguments{
\item{vignettes}{also check all \code{rmd} and \code{rnw} files in the pkg vignettes folder}

\item{lang}{preferred spelling langage. Usually either \code{"en-US"} or \code{"en-GB"}}

\item{error}{should \code{CMD check} fail if spelling errors are found? Default only
prints results}
}
\description{
Adds a unit test which automatically runs a spell check on documentation and/or
vignettes using the \link[spelling:spell_check_package]{spelling} package.
}
