% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_template}
\alias{use_template}
\title{Use a usethis template}
\usage{
use_template(template, save_as = template, data = list(), ignore = FALSE,
  open = FALSE)
}
\arguments{
\item{template}{Template name}

\item{save_as}{Name of file to create. Defaults to \code{save_as}}

\item{data}{A list of data passed to the template.}

\item{ignore}{Should the newly created file be added to \code{.Rbuildignore?}}

\item{open}{Should the new created file be opened in RStudio?}
}
\value{
A logical vector indicating if file was modified.
}
\description{
Use a usethis template
}
\keyword{internal}
