% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{use_testthat}
\alias{use_testthat}
\alias{use_test}
\title{Create tests}
\usage{
use_testthat()

use_test(name = NULL, open = interactive())
}
\arguments{
\item{name}{Base of test file name. If \code{NULL}, and you're using RStudio, will
be based on the name of the file open in the source editor.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
\code{use_testthat()} sets up testing infrastructure, creating \code{tests/testthat.R}
and \code{tests/testthat/}, and adding testthat as a Suggested package.
\code{use_test()} creates \code{tests/testthat/test-<name>.R} and opens it for editing.
}
\seealso{
The \href{https://r-pkgs.org/tests.html}{testing chapter} of \href{https://r-pkgs.org}{R Packages}.
}
