% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{use_testthat}
\alias{use_test}
\alias{use_testthat}
\title{Add testing infrastructor}
\usage{
use_testthat(pkg = ".")

use_test(name, pkg = ".")
}
\description{
\code{use_testthat} sets up testing infrastructure, creating
\file{tests/testthat.R} and \file{tests/testthat/}, and
adding \pkg{testthat} to the suggested packages. \code{use_test}
creates \file{tests/testthat/test-<name>.R} and opens it for editing.
}

