% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{use_testthat}
\alias{use_test}
\alias{use_testthat}
\title{Create tests}
\usage{
use_testthat(base_path = ".")

use_test(name, base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}

\item{name}{Test name.}
}
\description{
\code{use_testthat} sets up testing infrastructure, creating
\file{tests/testthat.R} and \file{tests/testthat/}, and
adding \pkg{testthat} to the suggested packages. \code{use_test}
creates \file{tests/testthat/test-<name>.R} and opens it for editing.
}

