% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{use_testthat}
\alias{use_testthat}
\alias{use_test}
\title{Create tests}
\usage{
use_testthat()

use_test(name = NULL, open = interactive())
}
\arguments{
\item{name}{Test name. if \code{NULL}, and you're using RStudio, will use
the name of the file open in the source editor.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
\code{use_testthat()} sets up testing infrastructure, creating
\file{tests/testthat.R} and \file{tests/testthat/}, and
adding \pkg{testthat} to the suggested packages. \code{use_test()}
creates \file{tests/testthat/test-<name>.R} and opens it for editing.
}
