% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{use_testthat}
\alias{use_testthat}
\title{Sets up overall testing infrastructure}
\usage{
use_testthat()
}
\description{
Creates \code{tests/testthat/}, \code{tests/testthat.R}, and adds the testthat package
to the Suggests field. Learn more in \url{https://r-pkgs.org/tests.html}
}
\examples{
\dontrun{
use_testthat()

use_test()

use_test("something-management")
}
}
\seealso{
\code{\link[=use_test]{use_test()}} to create individual test files
}
