% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{use_tibble}
\alias{use_tibble}
\title{Use tidyverse tibbles in your package}
\usage{
use_tibble()
}
\description{
Does setup necessary to deal in objects of type \code{tbl_df} or "tibble" from
your package. The tibble is the tidyverse's variant of the \code{data.frame}.
\code{use_tibble()} ensures that tibbles returned by your package will use the
tibble print method. Specifically, this function:
\itemize{
\item Adds tibble to "Imports" in DESCRIPTION
\item Creates \code{R/utils-tibble.R} with the necessary roxygen template
}
}
