% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{use_tidy_eval}
\alias{use_tidy_eval}
\title{Use tidy evaluation}
\usage{
use_tidy_eval()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function is defunct because there's no longer a need to systemetically
import and re-export large numbers of functions in order to use tidy
evaluation. Instead, use \code{\link[=use_import_from]{use_import_from()}} to tactically import functions
as you need them.
}
\keyword{internal}
