% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{use_tidy_thanks}
\alias{use_tidy_thanks}
\title{Identify contributors via GitHub activity}
\usage{
use_tidy_thanks(owner = NULL, repo = NULL, since = NULL)
}
\arguments{
\item{owner}{Name of user or organisation who owns the repo. Default is to
infer from Git remotes of active project.}

\item{repo}{Repository name, usually same as package name. Default is to
infer from Git remotes of active project.}

\item{since}{Timestamp in ISO 8601 format, passed along to the \href{https://developer.github.com/v3/issues/#list-issues-for-a-repository}{GitHub API endpoint for listing issues}.
Default is to use timestamp of commit associated with most recent GitHub
release.}
}
\value{
A character vector of GitHub usernames, invisibly.
}
\description{
Derives a list of GitHub usernames, based on who has opened issues or pull
requests. Used to populate the acknowledgment section of package release blog
posts at \url{https://www.tidyverse.org/articles/}. All arguments can potentially
be determined from the active project, if the project follows standard
practices around the GitHub remote and GitHub releases.
}
\examples{
\dontrun{
use_tidy_thanks()
use_tidy_thanks(owner = "r-lib", repo = "usethis")
use_tidy_thanks(owner = "r-lib", repo = "usethis", since = "2018-05-01")
use_tidy_thanks(since = "2018-02-24T00:13:45Z")
}
}
