% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{use_tutorial}
\alias{use_tutorial}
\title{Create a tutorial.}
\usage{
use_tutorial(name, title, open = interactive())
}
\arguments{
\item{name}{Base for file name to use for new tutorials. Should consist only
of numbers, letters, _ and -. We recommend using lower case.}

\item{title}{The title of the tutorial}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}
}
\description{
Creates new tutorial in \code{inst/tutorials}. Tutorials are interactive R Markdown documents built with the \code{learnr} package.
}
\section{General setup}{

\itemize{
\item Adds needed packages to \code{DESCRIPTION}.
\item Adds \code{inst/tutorials/*.html} to \code{.gitignore} so
you never accidental track rendered tutorials.
}
}

\examples{
\dontrun{
use_tutorial("learn-to-do-stuff", "Learn to do stuff")
}
}
\seealso{
The \href{https://rstudio.github.io/learnr/index.html}{learnr package documentation}.
}
