% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-upkeep.R
\name{use_upkeep_issue}
\alias{use_upkeep_issue}
\title{Open an issue in GitHub with a checklist of upkeep tasks}
\usage{
use_upkeep_issue(year = NULL)
}
\arguments{
\item{year}{Optional year you are performing the upkeep}
}
\description{
Open an issue in GitHub with a checklist of upkeep tasks
}
\examples{
\dontrun{
use_upkeep_issue(2023)
}
}
