% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{use_version}
\alias{use_version}
\title{Increment version number}
\usage{
use_version(which = NULL)
}
\arguments{
\item{which}{Which level to increment, must be one of "dev", "patch",
"minor" or "major". Defaults to NULL which activates interactive
control.}
}
\description{
This increments the "Version" field in \code{DESCRIPTION}, adds a new heading
to \code{NEWS.md} (if it exists), and then checks the result into git.
}
