% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{use_version}
\alias{use_version}
\title{Increment version number}
\usage{
use_version(level = NULL, increment = 1)
}
\arguments{
\item{level}{Which level to increment, must be one of "Mayor", "Minor",
"Patch" or "Development". Defaults to NULL which activates interactive
control.}

\item{increment}{Numerical. How much to increment the level by.
Defaults to 1.}
}
\description{
This increments the "Version" field in \code{DESCRIPTION}, adds a new heading
to \code{NEWS.md} (if it exists), and then checks the result into git.
}
