% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{use_vignette}
\alias{use_vignette}
\title{Create a vignette}
\usage{
use_vignette(name)
}
\arguments{
\item{name}{File name to use for new vignette. Should consist only of
numbers, letters, _ and -. I recommend using lower case.}
}
\description{
Adds needed packages to \code{DESCRIPTION}, and creates draft vignette
in \code{vignettes/}. It adds \code{inst/doc} to \code{.gitignore}
so you don't accidentally check in the built vignettes.
}
