% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{use_vignette}
\alias{use_vignette}
\title{Create a vignette}
\usage{
use_vignette(name)
}
\arguments{
\item{name}{Base for file name to use for new vignette. Should consist only
of numbers, letters, _ and -. I recommend using lower case.}
}
\description{
Performs general set up for vignettes and initializes a new individual
vignette:
\itemize{
\item Adds needed packages to \code{DESCRIPTION}
\item Adds \code{inst/doc} to \code{.gitignore} so built vignettes aren't tracked
\item Creates a new draft \code{.Rmd} vignette in \code{vignettes/} and, if possible,
opens it for editing
}
}
\examples{
\dontrun{
use_vignette("how-to-do-stuff")
}
}
\seealso{
The \href{http://r-pkgs.had.co.nz/vignettes.html}{vignettes chapter} of
\href{http://r-pkgs.had.co.nz}{R Packages}.
}
