% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{usethis-defunct}
\alias{usethis-defunct}
\alias{pr_pull_upstream}
\alias{browse_github_token}
\alias{browse_github_pat}
\alias{github_token}
\title{Defunct and deprecated functions in usethis}
\usage{
pr_pull_upstream()

browse_github_token(...)

browse_github_pat(...)

github_token()
}
\description{
These functions have either been deprecated or removed from usethis.
}
\section{\code{pr_pull_upstream()}}{

This function has been replaced by \code{\link[=pr_merge_main]{pr_merge_main()}}.
}

\section{\code{browse_github_token()}, \code{browse_github_pat()}}{

These functions have been replaced by \code{\link[=create_github_token]{create_github_token()}}.
}

\keyword{internal}
