% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-defunct.R
\name{usethis-defunct}
\alias{usethis-defunct}
\alias{pr_pull_upstream}
\alias{pr_sync}
\alias{browse_github_token}
\alias{browse_github_pat}
\alias{github_token}
\title{Defunct and deprecated functions in usethis}
\usage{
pr_pull_upstream()

pr_sync()

browse_github_token(...)

browse_github_pat(...)

github_token()
}
\description{
These functions have either been deprecated or removed from usethis.
}
\section{\code{pr_pull_upstream()}}{

This function has been replaced by \code{\link[=pr_merge_main]{pr_merge_main()}}.
}

\section{\code{pr_sync()}}{

Bundling these operations together did not seem justified, in terms of how
rarely this comes up and, when it does, how likely merge conflicts are.
Users of \code{pr_sync()} should implement these steps "by hand":
\itemize{
\item (Check you are on a PR branch)
\item \code{pr_pull()}
\item \code{pr_merge_main()}, deal with any merge conflicts, if any
\item \code{pr_push()}
}
}

\section{\code{browse_github_token()}, \code{browse_github_pat()}}{

These functions have been replaced by \code{\link[=create_github_token]{create_github_token()}}.
}

\section{\code{github_token()}}{

All implicit and explicit token discovery routes through \code{\link[gh:gh_token]{gh::gh_token()}}
now.
}

\keyword{internal}
