% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis-package.R
\docType{package}
\name{usethis-package}
\alias{usethis}
\alias{usethis-package}
\title{usethis: Automate Package and Project Setup}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Automate package and project setup tasks that are otherwise
    performed manually. This includes setting up unit testing, test 
    coverage, continuous integration, Git, 'GitHub', licenses, 'Rcpp', 'RStudio' 
    projects, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usethis.r-lib.org}
  \item \url{https://github.com/r-lib/usethis}
  \item Report bugs at \url{https://github.com/r-lib/usethis/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com} (0000-0003-4757-117X)

Authors:
\itemize{
  \item Jennifer Bryan \email{jenny@rstudio.com} (0000-0002-6983-2759)
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
